% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootbeta-methods.R
\name{bootbeta-methods}
\alias{bootbeta-methods}
\alias{plot.bootbeta}
\title{Methods for class "bootbeta"}
\usage{
\method{plot}{bootbeta}(x, ...)
}
\arguments{
\item{x}{A bootbeta object.}

\item{...}{Unused optional arguments.}
}
\description{
Method for plotting bootbeta objects returned by \code{\link[=bootbeta]{bootbeta()}}.
}
