\name{bootstrap}
\alias{bootstrap}
\title{Parametric bootstrapping}
\description{TODO}
%\details{}
\usage{
bootstrap(r, series, constants, ...)
}
\arguments{
\item{r}{the number of replications.}
\item{series, constants}{.}
\item{\dots}{optional arguments passed to \code{\link{sir}},
  used to simulate new time series data.}
}
%\value{}
%\seealso{}
%\references{}
\examples{
## TODO
}
