% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsbeta-methods.R
\name{bsbeta-methods}
\alias{bsbeta-methods}
\alias{plot.bsbeta}
\title{Methods for class "bsbeta"}
\usage{
\method{plot}{bsbeta}(x, ...)
}
\arguments{
\item{x}{A bsbeta object.}

\item{...}{Unused optional arguments.}
}
\description{
Method for plotting bsbeta objects returned by \code{\link[=bsbeta]{bsbeta()}}.
}
