% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bsbeta.R
\name{bsbeta}
\alias{bsbeta}
\title{\loadmathjax
Bootstrap CIs on transmission rate estimates}
\usage{
bsbeta(x, y = NULL, n = 100, p = 1, delay_dist = c(1))
}
\arguments{
\item{x}{A fastbeta object defining a \code{\link[=fastbeta]{fastbeta()}} estimate
of \mjseqn{\beta(t)}. Must satisfy \code{x$method \%in\% c("si", "sei")}.}

\item{y}{An optional loess object defining a loess fit to the
\code{\link[=fastbeta]{fastbeta()}} estimate of \mjseqn{\beta(t)} (see Details).
\code{\link[=try_loess]{try_loess()}} can help define a reasonable value for \code{y}
(see Examples).}

\item{n}{An integer scalar. The desired number of bootstrap
estimates of \mjseqn{\beta(t)}.}

\item{p}{A numeric vector of length \code{nrow(x$out)}. \code{p[i]} is the
positive probability that an infection between times \code{x$out$t[i-1]}
and \code{x$out$t[i]} is eventually reported. Alternatively, a numeric
scalar indicating a constant positive probability.}

\item{delay_dist}{A numeric vector. The distribution of the integer
number of observation intervals between infection and reporting.
\code{delay_dist[i]} is the probability that an infection that is
eventually reported is reported after \code{i-1} observation intervals.
\code{delay_dist} is replaced with \code{delay_dist / sum(delay_dist)} in
the event that \code{sum(delay_dist) != 1}.}
}
\value{
A bsbeta object. A list with elements:

\describe{
\item{\code{times}}{A numeric vector listing time points for the
bootstrap estimates of \mjseqn{\beta(t)}. Equal to \code{x$out$t}.
}
\item{\code{beta}}{A numeric vector. If \code{y = NULL}, the \code{\link[=fastbeta]{fastbeta()}}
estimate of \mjseqn{\beta(t)}, equal to \code{x$out$beta}. Otherwise,
the loess fit evaluated at \code{times}, equal to \code{predict(y, times)}.
}
\item{\code{mat}}{A numeric matrix whose \code{j}th column is the
\code{j}th bootstrap estimate of \mjseqn{\beta(t)}.
}
\item{\code{ci95}}{A numeric matrix with 2 columns giving the lower
and upper endpoints of the bootstrap 95\% confidence interval.
Equal to
\code{t(apply(mat, 1, function(x) if (all(is.finite(x))) quantile(x, probs = c(0.025, 0.975)) else rep(NA, 2)))}.
}
\item{\code{mean}}{A numeric vector. The mean bootstrap estimate.
Equal to \code{rowMeans(mat)}.
}
\item{\code{sd}}{A numeric vector. The standard deviation of
the bootstrap estimates. Equal to \code{apply(mat, 1, sd)}.
}
\item{\code{n}}{An integer scalar. The number of bootstrap estimates.
Equal to \code{ncol(mat)}.
}
\item{\code{call}}{The function call. The bsbeta object is reproducible
with \code{eval(call)}.
}
\item{\code{arg_list}}{A list of the arguments in the function call.
The bsbeta object is reproducible with \code{do.call(bsbeta, arg_list)}.
}
}
}
\description{
Constructs bootstrap 95\% confidence intervals on \code{\link[=fastbeta]{fastbeta()}} estimates
of a time-varying transmission rate \mjseqn{\beta(t)} (or alternatively
on smooth loess fits to potentially noisy \code{\link[=fastbeta]{fastbeta()}} estimates).
}
\details{
The \code{\link[=fastbeta]{fastbeta()}} estimate of \mjseqn{\beta(t)} (or a smooth loess fit
to the estimate) is taken to be the true transmission rate for the
purpose of simulating new time series data using \code{\link[=make_data]{make_data()}} and
generating bootstrap estimates of \mjseqn{\beta(t)} from those data
using \code{\link[=fastbeta]{fastbeta()}}. The bootstrap 95\% CI is then defined as the
2.5-97.5th percentile interval of the (unsmoothed) bootstrap estimates.

Arguments \code{p} and \code{delay_dist} can be used to define a model
for observation error in the data-generating process. Their
default values are such that all infections are reported with
no delay between infection and reporting, i.e., such that
there is no observation error. Deconvolution is employed to
reconstruct incidence from simulated reported incidence,
allowing for \mjseqn{\beta(t)} estimation using \code{\link[=fastbeta]{fastbeta()}}.
The iterative deconvolution algorithm is stopped when the
chi-squared criterion is satisfied or when the prescribed
maximum number of iterations (25) is reached (whichever occurs
first). See Value of \code{\link[=deconvol]{deconvol()}} with \code{simple = TRUE} for
details.
}
\examples{
# Simulate time series data using an SIR model
# with seasonally forced transmission rate
pl <- make_par_list(model = "sir")
df <- make_data(pl, with_ds = TRUE, model = "sir")

# Estimate the seasonally forced transmission rate
# using the SI method
fastbeta_out <- fastbeta(df, pl, method = "si")
plot(fastbeta_out)

# Fit a loess curve to the `fastbeta()` estimate.
# Try different values for the smoothing parameter
# and choose one that produces a reasonable fit.
try_loess_out <- try_loess(beta ~ t,
  data = fastbeta_out$out,
  q = seq(20, 90, by = 10),
  control = loess.control(surface = "direct"),
  plot = TRUE
)
my_loess <- try_loess_out[["q50"]] # `q = 50` seems reasonable

# Construct bootstrap 95\% confidence intervals
# on the loess estimate, conditional on a given
# observation model
bsbeta_out <- bsbeta(fastbeta_out, my_loess,
  n = 100,
  p = 1, # reporting probability: 50\%
  delay_dist = c(0, 0.5, 0.5) # reporting delay: 1 dt or 2 dt
                              # with equal probability
)
plot(bsbeta_out)

}
\seealso{
\link[=bsbeta-methods]{methods for class "bsbeta"},
\code{\link[=fastbeta]{fastbeta()}}, \code{\link[=try_loess]{try_loess()}}, \code{\link[=make_data]{make_data()}}
}
