% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_rrmse.R
\name{compute_rrmse}
\alias{compute_rrmse}
\title{Compute RRMSE in a vector}
\usage{
compute_rrmse(x, xhat, na_rm = FALSE)
}
\arguments{
\item{x}{A numeric vector.}

\item{xhat}{A numeric vector of the same length as \code{x}.
An estimate of \code{x}.}

\item{na_rm}{A logical scalar. If \code{TRUE}, then missing values
are ignored.}
}
\value{
A numeric scalar indicating the RRMSE in \code{xhat}, with exceptions
(see Details).
}
\description{
Computes the relative root mean square error (RRMSE) in a vector.
}
\details{
Let \code{cc = is.finite(x) & is.finite(xhat)} ("complete cases"). \code{cc} is the
vector of indices \code{i} such that both \code{x[i]} and \code{xhat[i]} are finite
(not missing and not infinite). If \code{length(cc) = 0} or if \code{na_rm = FALSE}
and \code{length(cc) < length(x)}, then \code{compute_rrmse()} returns \code{NA}. Otherwise,
it returns \code{sqrt(sum((xhat[cc] - x[cc])^2) / length(cc)) / mean(x[cc])}.
This is the RRMSE in \code{xhat[cc]} unless \code{mean(x[cc]) = 0}, in which case
RRMSE is not defined and the expression is evaluated accordingly as \code{Inf}
or \code{NaN}.
}
\examples{
x <- rep(10, 100)
xhat <- x + rnorm(x, 0, 1)
compute_rrmse(x, xhat) # RRMSE in `xhat`

}
