% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convol-methods.R
\name{convol-methods}
\alias{convol-methods}
\alias{plot.convol}
\title{Methods for class "convol"}
\usage{
\method{plot}{convol}(x, ...)
}
\arguments{
\item{x}{A convol object.}

\item{...}{Unused optional arguments.}
}
\description{
Method for plotting convol objects returned by \code{\link[=convol]{convol()}}.
}
