% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convol.R
\name{convol}
\alias{convol}
\title{Convolve an incidence time series with a delay distribution}
\usage{
convol(x, delay_dist = c(1), n = 0L)
}
\arguments{
\item{x}{An integer vector defining an equally spaced incidence time series.
Lists the number of infections on each day from day 0 to day \code{length(x)-1}.}

\item{delay_dist}{A numeric vector. The distribution of the number of days
from infection to reporting. \code{delay_dist[i]} is the probability that
an infection is reported after \code{i-1} days. \code{delay_dist} is replaced with
\code{delay_dist / sum(delay_dist)} in the event that \code{sum(delay_dist) != 1}.}

\item{n}{A non-negative integer scalar. The number of simulations desired.}
}
\value{
A convol object. A list with elements:

\describe{
\item{\code{times}}{A numeric vector equal to \code{0:(length(x)-1+b)},
where \code{b = max(which(delay_dist > 0)) - 1} is the maximum number
of days from infection to reporting, giving time points in days
for the convolution of \code{x} and \code{delay_dist}.
}
\item{\code{x_pad}}{A numeric vector equal to \code{c(x, rep(NA, b))}.
A convenience allowing one to plot the incidence time series
with \code{lines(times, x_pad, ...)}.
}
\item{\code{convolution}}{A numeric vector of length \code{length(times)}
giving the convolution of \code{x} and \code{delay_dist}. \code{convolution[i]}
is the expected number of infections reported on each day \code{times[i]}.
}
\item{\code{simulation}}{If \code{n = 0}, then \code{NULL}. Otherwise,
a numeric matrix with \code{length(times)} rows and \code{n} columns.
Each column is the result of sampling \code{delay_dist} \code{sum(x)} times
and binning the \code{sum(x)} infections counted in \code{x} forward in time
according to the sampled delays. \code{simulation[i, j]} is the number
of infections reported on day \code{times[i]} in simulation \code{j} of \code{n}.
}
\item{\code{call}}{The function call. The convol object is reproducible
with \code{eval(call)}.
}
\item{\code{arg_list}}{A list of the arguments in the function call. The
convol object is reproducible with \code{do.call(convol, arg_list)}.
}
}
}
\description{
Convolves an equally spaced incidence time series with a discrete
distribution of the number of days from infection to reporting,
generating an expected reported incidence time series. In addition,
simulates reported incidence by sampling from the delay distribution.
}
\details{
The algorithm does not explicitly require daily observations. For
example, \code{x} can be described most generally as listing the number
of infections during each observation interval from interval 0 to
interval \code{length(x)-1}. A different time step (e.g., one week instead
of one day) can be chosen as long as the choice is consistent between
\code{x} and \code{delay_dist}. The output must be interpreted accordingly.
}
\examples{
x <- floor(400 * exp(-seq(-2, 2, by = 0.1)^2))
delay_dist <- rep(1 / 11, 11)
n <- 5
convol_out <- convol(x, delay_dist, n)
plot(convol_out)

}
\seealso{
\link[=convol-methods]{methods for class "convol"}, \code{\link[=deconvol]{deconvol()}}
}
