% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvol-methods.R
\name{deconvol-methods}
\alias{deconvol-methods}
\alias{plot.deconvol}
\title{Methods for class "deconvol"}
\usage{
\method{plot}{deconvol}(x, ...)
}
\arguments{
\item{x}{A deconvol object.}

\item{...}{Unused optional arguments.}
}
\description{
Method for plotting deconvol objects returned by \code{\link[=deconvol]{deconvol()}}.
}
