% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvol.R
\name{deconvol}
\alias{deconvol}
\title{Deconvolve a reported incidence time series}
\usage{
deconvol(x, delay_dist = c(1), p = 1, it_max = 20L)
}
\arguments{
\item{x}{A numeric vector defining an equally spaced reported incidence
time series. Lists the number of cases reported on each day from day 0
to day \code{length(x)-1}.}

\item{delay_dist}{A numeric vector. The distribution of the number of days
from infection to reporting. \code{delay_dist[i]} is the probability that an
infection that is eventually reported is reported after \code{i-1} days.
\code{delay_dist} is replaced with \code{delay_dist / sum(delay_dist)} in the
event that \code{sum(delay_dist) != 1}.}

\item{p}{Either
(i) a numeric scalar giving a constant positive probability that an
infection is reported, or
(ii) a numeric vector giving a positive probability for each day
from day \code{-b} to day \code{length(x)-1},
where \code{b = max(which(delay_dist > 0)) - 1} is the maximum number of
days from infection to reporting. In the latter case, \code{p[i]} is the
probability that an infection on day \code{i-1-b} is reported.}

\item{it_max}{A non-negative integer. The desired number of iterations.}
}
\value{
A deconvol object. A list with elements:

\describe{
\item{\code{times}}{A numeric vector equal to \code{-b:(length(x)-1)},
where \code{b = max(which(delay_dist > 0)) - 1} is the maximum number
of days from infection to reporting, giving time points in days
for deconvolved incidence.
}
\item{\code{x_pad}}{A numeric vector equal to \code{c(rep(NA, b), x)}.
A convenience allowing one to plot the observed reported incidence
time series with \code{lines(times, x_pad, ...)}.
}
\item{\code{poisson}}{A numeric matrix. \code{poisson[i, j]} is the estimate
after \code{j-1} iterations of the expected number of infections on day
\code{times[i]} that are eventually reported. See Details for the initial
estimate.
}
\item{\code{inc}}{A numeric matrix. \code{inc[i, j]} is the expected number
of infections on day \code{times[i]} conditional on \code{poisson[, j]}.
\code{inc[, j]} gives deconvolved incidence after \code{j-1} iterations.
}
\item{\code{inc_rep}}{A numeric matrix. \code{inc_rep[i, j]} is the expected
number of cases reported on day \code{times[i]} conditional on \verb{poisson[, j].}
}
\item{\code{chi2}}{A numeric vector. \code{chi2[j]} is the normalized chi-squared
statistic corresponding to \code{inc_rep[, j]}, computed as
\code{mean((x - y)^2 / y)}, where \code{y = inc_rep[b+1:length(x), j]}.
}
\item{\code{it_chi2_lt1}}{An integer scalar equal to \code{min(which(chi2 < 1)) - 1}
if \code{any(chi2 < 1)} is \code{TRUE} and \code{NA} otherwise, giving the number
of iterations performed before the chi-squared criterion suggested
by \insertCite{Gold+09;textual}{fastbeta} (namely \code{chi2 < 1}) was
satisfied. Note that the corresponding incidence time series is
\code{inc[, 1+it_chi2_lt1]}.
}
\item{\code{call}}{The function call. The deconvol object is reproducible
with \code{eval(call)}.
}
\item{\code{arg_list}}{A list of the arguments in the function call. The
deconvol object is reproducible with \code{do.call(deconvol, arg_list)}.
}
}

Note that \code{poisson}, \code{inc}, and \code{inc_rep} all have \code{b+length(x)} rows
(corresponding to the elements of \code{times}) and \code{1+it_max} columns
(one for each iteration, starting from an initial value).
}
\description{
Deconvolves an equally spaced reported incidence time series using a
Richardson-Lucy iteration, generating an estimate of the underlying
incidence time series.
}
\section{Details}{
\subsection{1. Differences compared to \insertCite{Gold+09;textual}{fastbeta}}{

The algorithm follows \insertCite{Gold+09;textual}{fastbeta} with the
following differences:
\itemize{
\item Reported incidence is observed from day 0 to day \code{length(x)-1},
not from day 1 to day \code{length(x)}.
\item Infection and reporting can happen on the same day. A delay
of zero days can be assigned nonzero probability by setting
\code{delay_dist[1] > 0}.
\item The initial value for the Richardson-Lucy iteration is obtained by
binning all \code{sum(x)} cases between day \code{-b} and day \code{length(x)-1},
where \code{b = max(which(delay_dist > 0)) - 1} is the maximum number
of days from infection to reporting. Binning follows \code{delay_dist},
so that \code{delay_dist[i]} is the fraction of \code{x[j]} placed in the
bin for day \code{j-i}. This differs from
\insertCite{Gold+09;textual}{fastbeta}, whose initial value is \code{x}
shifted backwards in time by \code{s = which.max(delay_dist) - 1} days.
An undesired consequence of their choice is that deconvolved
incidence is necessarily zero from day \code{-b} to day \code{-s-1} and
from day \code{length(x)-s} to day \code{length(x)-1}.
}
}

\subsection{2. Noise in deconvolved incidence and choosing when to stop}{

\insertCite{Gold+09;textual}{fastbeta} suggest to stop iterating
once \code{chi2 < 1} and to keep the last generated incidence time series
(see Value, under \code{chi2} and \code{it_chi2_lt1}). The reason for this
criterion is that deconvolved incidence tends to acquire undesired
noise after several iterations, and it is typically better to stop
before this happens (see the supplement to
\insertCite{Gold+09;textual}{fastbeta} for details). However, this
criterion is not guaranteed to produce an optimal result, hence a
decision about which incidence time series to keep should be based
on a graphical exploration of the entire \code{deconvol()} output
(see Examples).
}

\subsection{3. Missing values in output}{

The last \code{a} rows of \code{poisson} and \code{inc} are \code{NA}, where
\code{a = min(which(delay_dist > 0)) - 1} is the minimum number of days from
infection to reporting. For \code{a > 0}, the probability that an infection
between day \code{length(x)-a} and day \code{length(x)-1} is reported between day
\code{0} and day \code{length(x)-1} is zero, so the supplied reported incidence
time series provides no information about incidence on those \code{a} days.
}

\subsection{4. Observation interval}{

The algorithm does not explicitly require daily observations.
For example, \code{x} can be described most generally as listing the
number of infections reported during each observation interval
from interval 0 to interval \code{length(x)-1}. A different observation
interval (e.g., one week instead of one day) can be chosen as long
as the choice is consistent across \code{x}, \code{delay_dist}, and \code{p}.
The output must be interpreted accordingly.
}
}

\examples{
x <- 400 * exp(-seq(-2, 2, by = 0.1)^2)
delay_dist <- rep(1 / 11, 11)
p <- 0.5
it_max <- 25
dcv_out <- deconvol(x, delay_dist, p, it_max)
plot(dcv_out) # creates 3 plots

}
\references{
\insertRef{Gold+09}{fastbeta}
}
\seealso{
\link[=deconvol-methods]{methods for class "deconvol"}, \code{\link[=convol]{convol()}}
}
