% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolve.R
\name{deconvolve}
\alias{deconvolve}
\title{\loadmathjax
Reconstruct incidence from reported incidence}
\usage{
deconvolve(
  x,
  delay_dist = c(1),
  p = rep(1, length(x) + length(delay_dist) - 1),
  max_it = 20L
)
}
\arguments{
\item{x}{Reported incidence time series. A numeric vector giving the
number of cases reported on each day from day \code{0} to day \code{length(x)-1}.}

\item{delay_dist}{Distribution of the number of days from infection to
reporting. A numeric vector. \code{delay_dist[i]} is the probability that
an infection that is eventually reported is reported after \code{i-1} days.
An error is thrown if \code{sum(delay_dist) != 1}.}

\item{p}{Reporting probability. Either
(i) a numeric scalar giving a positive probability that an infection
is reported, or
(ii) a numeric vector giving a positive probability for each day
from day \code{-b} to day \code{length(x)-1},
where \code{b = max(which(delay_dist > 0)) - 1} is the maximum number of
days from infection to reporting. In the latter case, \code{p[i]} is the
probability that an infection on day \code{i-1-b} is reported.}

\item{max_it}{Maximum number of iterations. A non-negative integer.}
}
\value{
A numeric vector of length \code{b+length(x)} giving an estimate of the
number of infections on each day from day \code{-b} to day \code{length(x)-1},
where \code{b = max(which(delay_dist > 0)) - 1} is the maximum number of
days from infection to reporting. It is the last Richardson-Lucy
iterate, scaled by \code{1 / p}.

The returned vector has attributes \code{call} and \code{arg_list}, making it
reproducible with \code{eval(call)} or \code{do.call(deconvolve, arg_list)}.
It has a third attribute \code{mat}, a matrix containing the full output
of the Richardson-Lucy iteration, prior to scaling by \code{1 / p}. There
is one row for each day from day \code{-b} to day \code{length(x)-1} and one
column for each iteration (not counting the first column, which gives
the initial value for the iteration). Hence the returned vector is
the last column of \code{mat}, scaled by \code{1 / p}.
}
\description{
\code{deconvolve()} employs a Richardson-Lucy iteration to deconvolve a reported
incidence time series. The supplied delay distribution is used to carry out
the deconvolution, and the result is scaled by the supplied
reporting probability to generate an incidence time series.
}
\references{
\insertRef{Gold+09}{fastbeta}
}
