% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_beta.R
\name{estimate_beta}
\alias{estimate_beta}
\alias{estimate_beta_fc}
\alias{estimate_beta_s}
\alias{estimate_beta_si}
\title{\loadmathjax
Estimate time-varying transmission rates}
\usage{
estimate_beta_fc(df, par_list)

estimate_beta_s(df, par_list)

estimate_beta_si(df, par_list)
}
\arguments{
\item{df}{A data frame with numeric columns:

\describe{
\item{\code{t}}{Time. \code{t[i]} is equal to
\mjseqn{t_{i-1} = t_0 + (i-1) \Delta t} in any convenient units,
where \mjseqn{t_0} is the initial observation time
and \mjseqn{\Delta t} is the observation interval.
}
\item{\code{Z}}{Incidence. \code{Z[i]} is the number of infections between
times \code{t[i-1]} and \code{t[i]}.
}
\item{\code{B}}{Births. \code{B[i]} is the number of births between times
\code{t[i-1]} and \code{t[i]}.
}
\item{\code{mu}}{Natural mortality rate. \code{mu[i]} is the rate at time
\code{t[i]} expressed per unit \mjseqn{\Delta t} and per capita.
S and SI methods only.
}
}}

\item{par_list}{A list with elements:

\describe{
\item{\code{tgen}}{Mean generation interval of the disease of interest
in units of the observation interval \mjseqn{\Delta t}.
}
\item{\code{S0}}{Number of susceptibles at the initial observation time
\mjseqn{t_0}.
}
\item{\code{I0}}{Number of infecteds at the initial observation time
\mjseqn{t_0}. SI method only.
}
}}
}
\value{
A data frame with numeric columns:

\describe{
\item{\code{t}}{Time. Identical to \code{df$t}.}
\item{\code{Z}}{Incidence. Identical to \code{df$Z}.}
\item{\code{Z_agg}}{Aggregated incidence. \code{Z_agg[i]} is the number
of infections between times \code{t[i-round(tgen)+1]} and \code{t[i]},
for \code{i} in \code{seq(1 + round(tgen), nrow(df), by = round(tgen))}.
\code{Z_agg[i]} is \code{NA} otherwise. FC method only.
}
\item{\code{B}}{Births. Identical to \code{df$B}.}
\item{\code{B_agg}}{Aggregated births. \code{B_agg[i]} is the number
of births between times \code{t[i-round(tgen)+1]} and \code{t[i]},
for \code{i} in \code{seq(1 + round(tgen), nrow(df), by = round(tgen))}.
\code{B_agg[i]} is \code{NA} otherwise. FC method only.
}
\item{\code{mu}}{Natural mortality rate. Identical to \code{df$mu}.
S and SI methods only.
}
\item{\code{S}}{Number of susceptibles. \code{S[i]} is the estimated
number of susceptibles at time \code{t[i]}.
}
\item{\code{I}}{Number of infecteds. \code{I[i]} is the estimated
number of infecteds at time \code{t[i]}.
}
\item{\code{beta}}{Transmission rate. \code{beta[i]} is the estimated
transmission rate at time \code{t[i]}, expressed per unit \mjseqn{\Delta t}
per susceptible per infected.
}
}

The data frame has attributes \code{call} and \code{arg_list}, making it
reproducible with \code{eval(call)} or \code{do.call(estimate_beta_x, arg_list)}.
}
\description{
Functions implementing the FC, S, and SI methods (see Algorithm)
for estimating time-varying transmission rates \mjseqn{\beta(t)}
from time series of incidence, births, and natural mortality with
observations at equally spaced time points
\mjseqn{t_i = t_0 + i \Delta t}.
The FC and S methods are deprecated and outperformed by the
more robust SI method. \code{estimate_beta_si()} need not be called
directly: it is wrapped in the more useful constructor function
\code{\link[=fastbeta]{fastbeta()}}, which does checks on input and output and returns
a fastbeta object with associated methods.
}
\section{Details}{
\subsection{1. Algorithm}{
\subsection{FC method}{

The supplied incidence time series \mjseqn{Z_i} (\code{df$Z}) is aggregated
over generation intervals:

\mjsdeqn{Z_i^\text{agg} = \sum_{k=i-g+1}^k Z_k\,,\quad i=jg\,,\quad j = 1,2,\ldots\,,}

where \mjseqn{g = \mathrm{nint}(t_\text{gen} / \Delta t)} is the mean
generation interval \mjseqn{t_\text{gen}} in units of the observation
interval \mjseqn{\Delta t} (\code{par_list$tgen}), rounded to the nearest
integer. The supplied births time series \mjseqn{B_i} (\code{df$B})
is aggregated similarly. The susceptible population size is estimated
recursively starting from the supplied initial value \mjseqn{S_0}
(\code{par_list$S0}):

\mjsdeqn{S_{i+g} = S_i + B_{i+g}^\text{agg} - Z_{i+g}^\text{agg}\,,\quad i=jg\,,\quad j = 0,1,\ldots\,,}

where natural mortality of susceptibles is assumed to be negligible.
The infected population size is approximated by aggregated incidence:

\mjsdeqn{I_i = Z_i^\text{agg}\,,\quad i=jg\,,\quad j = 1,2,\ldots\,,}

where natural mortality of infecteds is assumed to be negligible.
The transmission rate is then estimated as

\mjsdeqn{\beta_i = \frac{Z_{i+g}}{S_i I_i g \Delta t}\,,\quad i=jg\,,\quad j = 1,2,\ldots\,.}

Note that this algorithm was first published by
\insertCite{FineClar82;textual}{fastbeta}.
}

\subsection{S method}{

The susceptible population size is estimated recursively starting from
the supplied initial value \mjseqn{S_0} (\code{par_list$S0}):

\mjsdeqn{S_{i+1} = S_i + B_{i+1} - Z_{i+1} - \mu_i S_i \Delta t\,.}

The infected population size is approximated by a scaling of incidence:

\mjsdeqn{I_i = \frac{Z_{i-g+1}}{(\gamma + \mu_i) \Delta t}\,,\quad \gamma = 1 / t_\text{gen}\,,}

where \mjseqn{g = \mathrm{nint}(t_\text{gen} / \Delta t)} is the mean
generation interval \mjseqn{t_\text{gen}} in units of the observation
interval \mjseqn{\Delta t} (\code{par_list$tgen}), rounded to the nearest
integer. The transmission rate is then estimated as

\mjsdeqn{\beta_i = \frac{Z_{i+1}}{S_i I_i \Delta t}\,.}
}

\subsection{SI method}{

The susceptible and infected population sizes are estimated recursively
starting from the supplied initial values \mjseqn{S_0} (\code{par_list$S0})
and \mjseqn{I_0} (\code{par_list$I0}):

\mjsdeqn{\begin{align*} S_{i+1} &= \frac{\big\lbrack 1 - \frac{1}{2} \mu_i \Delta t \big\rbrack S_i + B_{i+1} - Z_{i+1}}{1 + \frac{1}{2} \mu_{i+1} \Delta t}\,, \cr I_{i+1} &= \frac{\big\lbrack 1 - \frac{1}{2} (\gamma + \mu_i) \Delta t \big\rbrack I_i + Z_{i+1}}{1 + \frac{1}{2} (\gamma + \mu_{i+1}) \Delta t}\,,\quad \gamma = 1 / t_\text{gen}\,. \end{align*}}

The transmission rate is then estimated as

\mjsdeqn{\beta_i = \frac{Z_i + Z_{i+1}}{2 S_i I_i \Delta t}\,.}
}

}

\subsection{2. Missing data}{

Missing values in \code{df$Z} are not tolerated and must be imputed.
Columns \code{S} and \code{beta} in the output will be filled with \code{NA}
after the index of the first missing value.

In the FC and S methods, zeros in \code{df$Z} can cause divide-by-zero
errors. In this case, column \code{beta} in the output may contain \code{NaN}
and \code{Inf} where an estimate would otherwise be expected. In the SI
method, strings of zeros in \code{df$Z} can lead to spurious zeros and
large numeric elements in column \code{beta} of the output. Zeros in \code{df$Z}
can be "imputed" (replaced with positive values in a sensible way)
to avoid this issue.

Imputation is not carried out by \code{estimate_beta_fc()} and friends,
but \emph{is} carried out by the more useful constructor function
\code{\link[=fastbeta]{fastbeta()}}, a wrapper for \code{estimate_beta_si()}.
}

\subsection{3. Effective observation interval (FC method only)}{

In the absence of errors, every \code{round(par_list$tgen)}th row in
the FC method output will be complete. The remaining rows will
contain \code{NA} in columns \code{S}, \code{I}, \code{beta}, \code{Z_agg}, and \code{B_agg}.
This is not a mistake: the rounded generation interval becomes
the \emph{effective} observation interval when the incidence and births
time series are aggregated (see Algorithm).
}
}

\examples{
# Simulate time series data using an SIR model
# with seasonally forced transmission rate
par_list <- make_par_list(dt_weeks = 1)
df <- make_data(par_list = par_list, with_ds = FALSE)
head(df)

# Estimate susceptibles, infecteds, and
# the seasonally forced transmission rate
df_fc <- estimate_beta_fc(df, par_list)
head(df_fc)
df_s <- estimate_beta_s(df, par_list)
head(df_s)
df_si <- estimate_beta_si(df, par_list)
head(df_si)

# FC method fails rapidly as a result
# of ignoring susceptible mortality
plot(S ~ I(t - t[1]), df, type = "l", ylim = c(43, 83) * 1e03)
lines(S ~ I(t - t[1]), df_fc[!is.na(df_fc$S), ], col = "blue")
plot(I ~ I(t - t[1]), df, type = "l", ylim = c(0, 3) * 1e03)
lines(I ~ I(t - t[1]), df_fc[!is.na(df_fc$I), ], col = "blue")
plot(beta ~ I(t - t[1]), df, type = "l", ylim = c(0.5, 1.2) * 1e-05)
lines(beta ~ I(t - t[1]), df_fc[!is.na(df_fc$beta), ], col = "blue")

# S method does well
plot(S ~ I(t - t[1]), df, type = "l", ylim = c(43, 58) * 1e03)
lines(S ~ I(t - t[1]), df_s, col = "blue")
plot(I ~ I(t - t[1]), df, type = "l", ylim = c(0, 3) * 1e03)
lines(I ~ I(t - t[1]), df_s, col = "red")
plot(beta ~ I(t - t[1]), df, type = "l", ylim = c(0.95, 1.25) * 1e-05)
lines(beta ~ I(t - t[1]), df_s, col = "blue")

# SI method does well
plot(S ~ I(t - t[1]), df, type = "l", ylim = c(43, 58) * 1e03)
lines(S ~ I(t - t[1]), df_si, col = "blue")
plot(I ~ I(t - t[1]), df, type = "l", ylim = c(3, 3) * 1e03)
lines(I ~ I(t - t[1]), df_si, col = "blue")
plot(beta ~ I(t - t[1]), df, type = "l", ylim = c(0.95, 1.25) * 1e-05)
lines(beta ~ I(t - t[1]), df_si, col = "blue")

# However, restarting with `with_ds = TRUE`
# reveals that the S method is vulnerable
# to propagation of noise from the data to
# the transmission rate estimate. Hence the
# SI method should be preferred in practice.

}
\references{
\insertRef{Jaga+20}{fastbeta}
\insertRef{FineClar82}{fastbeta}
}
\seealso{
\code{\link[=fastbeta]{fastbeta()}}
}
