% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_beta_fc.R
\name{estimate_beta_fc}
\alias{estimate_beta_fc}
\title{\loadmathjax
Estimate time-varying transmission rates (FC method)}
\usage{
estimate_beta_fc(df = data.frame(), par_list = list())
}
\arguments{
\item{df}{A data frame with numeric columns:

\describe{
\item{\code{t}}{Time. \code{t[i]} is equal to
\mjseqn{t_{i-1} = t_0 + (i-1) \Delta t} in any convenient units,
where \mjseqn{t_0} is the initial observation time
and \mjseqn{\Delta t} is the observation interval.
}
\item{\code{Z}}{Incidence. \code{Z[i]} is the number of infections between
times \code{t[i-1]} and \code{t[i]}.
}
\item{\code{B}}{Births. \code{B[i]} is the number of births between times
\code{t[i-1]} and \code{t[i]}.
}
}}

\item{par_list}{A list with elements:

\describe{
\item{\code{S0}}{Number of susceptibles at the initial observation time
\mjseqn{t_0}.}
\item{\code{tgen}}{Mean generation interval of the disease of interest
in units of the observation interval \mjseqn{\Delta t}.
}
}}
}
\value{
A data frame with numeric columns:

\describe{
\item{\code{t}}{Time. Identical to \code{df$t}.}
\item{\code{Z}}{Incidence. Identical to \code{df$Z}.}
\item{\code{Z_agg}}{Aggregated incidence. \code{Z_agg[i]} is the number
of infections between times \code{t[i-round(tgen)+1]} and \code{t[i]},
for \code{i} in \code{seq(1 + round(tgen), nrow(df), by = round(tgen))}.
\code{Z_agg[i]} is \code{NA} otherwise.
}
\item{\code{B}}{Births. Identical to \code{df$B}.}
\item{\code{B_agg}}{Aggregated births. \code{B_agg[i]} is the number
of births between times \code{t[i-round(tgen)+1]} and \code{t[i]},
for \code{i} in \code{seq(1 + round(tgen), nrow(df), by = round(tgen))}.
\code{B_agg[i]} is \code{NA} otherwise.
}
\item{\code{S}}{Number of susceptibles. \code{S[i]} is the estimated
number of susceptibles at time \code{t[i]}.
}
\item{\code{I}}{Number of infecteds. \code{I[i]} is the estimated
number of infecteds at time \code{t[i]}.
}
\item{\code{beta}}{Transmission rate. \code{beta[i]} is the estimated
transmission rate at time \code{t[i]}, expressed per unit \mjseqn{\Delta t}
per susceptible per infected.
}
}

The data frame has attributes \code{call} and \code{arg_list}, making it
reproducible with \code{eval(call)} or \code{do.call(estimate_beta_fc, arg_list)}.
}
\description{
\code{estimate_beta_fc()} implements the FC method (see Algorithm) for
estimating time-varying transmission rates \mjseqn{\beta(t)} from
time series of incidence and births observed at equally spaced
time points \mjseqn{t_k = t_0 + k \Delta t}. The FC method is
deprecated and outperformed by the more robust SI method. See
\code{\link[=estimate_beta_si]{estimate_beta_si()}}.
}
\section{Details}{
\subsection{1. Algorithm}{

The supplied incidence time series \mjseqn{Z_k} (\code{df$Z}) is aggregated
over generation intervals:

\mjsdeqn{Z_k^\text{agg} = \sum_{i=k-g+1}^k Z_i\,,\quad k=jg\,,\quad j = 1,2,\ldots\,,}

where \mjseqn{g = \mathrm{nint}(t_\text{gen} / \Delta t)} is the mean
generation interval \mjseqn{t_\text{gen}} in units of the observation
interval \mjseqn{\Delta t} (\code{par_list$tgen}), rounded to the nearest
integer. The supplied births time series \mjseqn{B_k} (\code{df$B})
is aggregated similarly. The susceptible population size is estimated
recursively starting from the supplied initial value \mjseqn{S_0}
(\code{par_list$S0}):

\mjsdeqn{S_{k+g} = S_k + B_{k+g}^\text{agg} - Z_{k+g}^\text{agg}\,,\quad k=jg\,,\quad j = 0,1,\ldots\,,}

where natural mortality of susceptibles is assumed to be negligible.
The infected population size is approximated by aggregated incidence:

\mjsdeqn{I_k = Z_k^\text{agg}\,,\quad k=jg\,,\quad j = 1,2,\ldots\,,}

where natural mortality of infecteds is assumed to be negligible.
The transmission rate is then estimated as

\mjsdeqn{\beta_k = \frac{Z_{k+g}}{S_k I_k g \Delta t}\,,\quad k=jg\,,\quad j = 1,2,\ldots\,.}
}

\subsection{2. Missing data}{

Missing values in \code{df$Z} and \code{df$B} are not tolerated and must be
imputed. Columns \code{S} and \code{beta} in the output will be filled with
\code{NA} after the index of the first missing value.

Zeros in \code{df$Z} can cause divide-by-zero errors. In this case,
column \code{beta} in the output may contain \code{NaN} and \code{Inf} where
an estimate would otherwise be expected. Zeros in \code{df$Z} can be
"imputed" (replaced with positive values in a sensible way) to
avoid this issue.

Imputation is not carried out by \code{estimate_beta_fc()}.
}

\subsection{3. Effective observation interval}{

In the absence of errors, every \code{round(par_list$tgen)}th row in
the output will be complete. The remaining rows will contain \code{NA}
in columns \code{S}, \code{I}, \code{beta}, \code{Z_agg}, and \code{B_agg}. This is not a
bug: the (rounded) generation interval becomes the \emph{effective}
observation interval when the incidence and birth time series
are aggregated (see Algorithm).
}
}

\examples{
# Simulate time series of incidence and births using
# an SIR model with seasonally forced transmission rate
par_list <- make_par_list(dt_weeks = 1)
df <- make_data(
  par_list = par_list,
  n = 20 * 365 / 7, # 20 years is ~1042 weeks
  with_dem_stoch = FALSE
)
head(df)

# Estimate susceptibles, infecteds, and
# the seasonally forced transmission rate
df_fc <- estimate_beta_fc(df, par_list)
head(df_fc)

# Estimation of susceptibles and transmission rate fails
# because the FC method ignores susceptible mortality
plot(S ~ I(t - t[1]), df, type = "l", ylim = c(43, 83) * 1e03)
lines(S ~ I(t - t[1]), df_fc[!is.na(df_fc$S), ], col = "red")
plot(beta ~ I(t - t[1]), df, type = "l", ylim = c(0.5, 1.2) * 1e-05)
lines(beta ~ I(t - t[1]), df_fc[!is.na(df_fc$beta), ], col = "red")

}
\references{
\insertRef{deJo+20}{fastbeta}
}
