% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_beta_s.R
\name{estimate_beta_s}
\alias{estimate_beta_s}
\title{\loadmathjax
Estimate time-varying transmission rates (S method)}
\usage{
estimate_beta_s(df = data.frame(), par_list = list())
}
\arguments{
\item{df}{A data frame with numeric columns:

\describe{
\item{\code{t}}{Time. \code{t[i]} is equal to
\mjseqn{t_{i-1} = t_0 + (i-1) \Delta t} in any convenient units,
where \mjseqn{t_0} is the initial observation time
and \mjseqn{\Delta t} is the observation interval.
}
\item{\code{Z}}{Incidence. \code{Z[i]} is the number of infections between
times \code{t[i-1]} and \code{t[i]}.
}
\item{\code{B}}{Births. \code{B[i]} is the number of births between times
\code{t[i-1]} and \code{t[i]}.
}
\item{\code{mu}}{Natural mortality rate. \code{mu[i]} is the rate at time
\code{t[i]} expressed per unit \mjseqn{\Delta t} and per capita.
}
}}

\item{par_list}{A list with elements:

\describe{
\item{\code{S0}}{Number of susceptibles at the initial observation time
\mjseqn{t_0}.}
\item{\code{tgen}}{Mean generation interval of the disease of interest
in units of the observation interval \mjseqn{\Delta t}.
}
}}
}
\value{
A data frame with numeric columns:

\describe{
\item{\code{t}}{Time. Identical to \code{df$t}.}
\item{\code{Z}}{Incidence. Identical to \code{df$Z}.}
\item{\code{B}}{Births. Identical to \code{df$B}.}
\item{\code{mu}}{Natural mortality rate. Identical to \code{df$mu}.}
\item{\code{S}}{Number of susceptibles. \code{S[i]} is the estimated
number of susceptibles at time \code{t[i]}.
}
\item{\code{I}}{Number of infecteds. \code{I[i]} is the estimated
number of infecteds at time \code{t[i]}.
}
\item{\code{beta}}{Transmission rate. \code{beta[i]} is the estimated
transmission rate at time \code{t[i]}, expressed per unit \mjseqn{\Delta t}
per susceptible per infected.
}
}

The data frame has attributes \code{call} and \code{arg_list}, making it
reproducible with \code{eval(call)} or \code{do.call(estimate_beta_s, arg_list)}.
}
\description{
\code{estimate_beta_s()} implements the S method (see Algorithm) for
estimating time-varying transmission rates \mjseqn{\beta(t)} from
time series of incidence, births, and natural mortality observed
at equally spaced time points \mjseqn{t_k = t_0 + k \Delta t}.
The S method is deprecated and outperformed by the more robust SI
method. See \code{\link[=estimate_beta_si]{estimate_beta_si()}}.
}
\section{Details}{
\subsection{1. Algorithm}{

The susceptible population size is estimated recursively starting from
the supplied initial value \mjseqn{S_0} (\code{par_list$S0}):

\mjsdeqn{S_{k+1} = S_k + B_{k+1} - Z_{k+1} - \mu_k S_k \Delta t\,.}

The infected population size is approximated by a scaling of incidence:

\mjsdeqn{I_k = \frac{Z_{k-g+1}}{(\gamma + \mu_k) \Delta t}\,,\quad \gamma = 1 / t_\text{gen}\,,}

where \mjseqn{g = \mathrm{nint}(t_\text{gen} / \Delta t)} is the mean
generation interval \mjseqn{t_\text{gen}} in units of the observation
interval \mjseqn{\Delta t} (\code{par_list$tgen}), rounded to the nearest
integer. The transmission rate is then estimated as

\mjsdeqn{\beta_k = \frac{Z_{k+1}}{S_k I_k \Delta t}\,.}
}

\subsection{2. Missing data}{

Missing values in \code{df$Z}, \code{df$B}, \code{df$mu} are not tolerated and must
be imputed. Columns \code{S} and \code{beta} in the output will be filled with
\code{NA} after the index of the first missing value.

Zeros in \code{df$Z} can cause divide-by-zero errors. In this case,
column \code{beta} in the output may contain \code{NaN} and \code{Inf} where
an estimate would otherwise be expected. Zeros in \code{df$Z} can be
"imputed" (replaced with positive values in a sensible way) to
avoid this issue.

Imputation is not carried out by \code{estimate_beta_s()}.
}
}

\examples{
# Simulate time series of incidence and births using
# an SIR model with seasonally forced transmission rate
par_list <- make_par_list(dt_weeks = 1, epsilon = 0.5)
df <- make_data(
  par_list = par_list,
  n = 20 * 365 / 7, # 20 years is ~1042 weeks
  with_dem_stoch = TRUE
)
head(df)

# Estimate susceptibles, infecteds, and
# the seasonally forced transmission rate
df_s <- estimate_beta_s(df, par_list)
head(df_s)

# Inspect
plot(S ~ I(t - t[1]), df, type = "l", ylim = c(43, 58) * 1e03)
lines(S ~ I(t - t[1]), df_s, col = "red")
plot(I ~ I(t - t[1]), df, type = "l", ylim = c(0, 3) * 1e03)
lines(I ~ I(t - t[1]), df_s, col = "red")
plot(beta ~ I(t - t[1]), df, type = "l", ylim = c(0.95, 1.25) * 1e-05)
lines(beta ~ I(t - t[1]), df_s, col = "red")

}
\references{
\insertRef{deJo+20}{fastbeta}
}
