\name{fastbeta-deprecated}
\alias{fastbeta-deprecated}
\title{Deprecated Functions in Package \pkg{fastbeta}}
\description{
The functions and other objects listed here are provided only
for compatibility with older versions of \pkg{fastbeta} and may
become defunct as soon as the next release.
}
\usage{
## Nothing yet!
}
\seealso{
\code{\link{Defunct}}, \code{\link{base-defunct}},
\code{\link{fastbeta-defunct}}, \code{\link{fastbeta-notyet}}.
}
\keyword{internal}
