% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastbeta.R
\name{fastbeta-methods}
\alias{fastbeta-methods}
\alias{plot.fastbeta}
\alias{print.fastbeta}
\title{Methods for class fastbeta}
\usage{
\method{plot}{fastbeta}(x, ...)

\method{print}{fastbeta}(x, ...)
}
\arguments{
\item{x}{A fastbeta object.}

\item{...}{Unused optional arguments.}
}
\description{
Methods for printing and plotting fastbeta objects
returned by \code{\link[=fastbeta]{fastbeta()}}.
}
