\name{fastbeta-notyet}
\alias{fastbeta-notyet}
\title{Not Yet Implemented Functions in Package \pkg{fastbeta}}
\description{
The functions listed here are defined but not yet implemented.
Use \code{\link{bug.report}(package = "fastbeta")} to request
an implementation.
}
\usage{
## Nothing yet!
}
\seealso{
\code{\link{NotYetImplemented}},
\code{\link{fastbeta-deprecated}}, \code{\link{fastbeta-defunct}}.
}
\keyword{internal}
