% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_peak_times.R
\name{get_peak_times}
\alias{get_peak_times}
\title{Find times of peaks in periodic time series}
\usage{
get_peak_times(x, period, bw_mavg, bw_peakid)
}
\arguments{
\item{x}{Numeric vector. A periodic, equally spaced time series.
Time points are taken to be \code{seq_along(x)}.}

\item{period}{Numeric scalar. The period of \code{x} in units of the
observation interval.}

\item{bw_mavg}{Numeric scalar. Bandwidth for the moving average
applied to \code{x}. \code{x_mavg[i]} will be the mean of observations
\code{x[j]} with
\ifelse{latex}{\out{$|i - j| \leq bandwidth$}}{\ifelse{html}{\out{&vert;<i>i</i> &minus; <i>j</i>&vert; &le; bandwidth}}{|i - j| <= bandwidth}}.}

\item{bw_peakid}{Numeric scalar. Bandwidth for peak identification.
\code{i} will be a peak time if and only if \code{x_mavg[i] > x_mavg[j]}
for all \code{j} with
\ifelse{latex}{\out{$0 < |i - j| \leq bandwidth$}}{\ifelse{html}{\out{0 &lt; &vert;<i>i</i> &minus; <i>j</i>&vert; &le; bandwidth}}{0 < |i - j| <= bandwidth}}..}
}
\value{
A list containing:

\describe{
\item{\code{all}}{Numeric vector. A subset of \code{seq_along(x)}
listing the times of all identified peaks.
}
\item{\code{phase}}{Numeric vector. A subset of \code{seq_along(x)}
listing the times of all identified peaks in phase with the
first peak.
}
}

A list of the arguments of \code{get_peak_times()} is included as an
attribute.
}
\description{
\code{get_peak_times()} locates peaks in periodic, equally spaced time
series with known period. It applies a central moving average to the
raw time series, identifies when peaks occur in the resulting smooth
time series, and subsets those peaks in phase with the first peak.
}
\details{
The bandwidths \code{bw_mavg} and \code{bw_peakid} must be iteratively tuned
so as to disqualify peaks caused by noise in \code{x} (choose \code{bw_mavg}
large enough) without disqualifying true peaks (choose \code{bw_peakid}
not too large).

\code{get_peak_times()} will not detect a peak at time \code{i} if
\itemize{
\item \code{i <= floor(bw_mavg) + floor(bw_peakid)}, or
\item \code{i >= length(x) - floor(bw_mavg) - floor(bw_peakid) + 1}.
}

These peaks are systematically disqualified for two reasons:
\enumerate{
\item The moving average performed by \code{\link[stats:filter]{stats::filter()}} assigns
\code{x_mavg[i]} the value \code{NA} if there are any missing values
in the band centered at \code{x[i]}. Hence \code{x_mavg[i]} is always
\code{NA} if
\itemize{
\item \code{i <= floor(bw_mavg)} or
\item \code{i >= length(x) - floor(bw_mavg) + 1}.
}
\item Missing values in the band centered at \code{x_mavg[i]} disqualify
\code{i} as a peak time.
}

For example, if \code{bw_mavg = 1} and \code{bw_peakid = 2}, then \code{x_mavg[1]}
is assigned the value \code{NA}, and this disqualifies any peak at index
1, 2, or 3.
}
