% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impute_na.R
\name{impute_na}
\alias{impute_na}
\title{Impute missing values in a numeric vector}
\usage{
impute_na(x, zero_as_na = FALSE)
}
\arguments{
\item{x}{Numeric vector.}

\item{zero_as_na}{Logical scalar. If \code{TRUE},
then zeros in \code{x} are treated like missing values.}
}
\value{
\code{x} with missing values replaced where possible.
}
\description{
Imputes missing values in a numeric vector
by linear interpolation of observed values,
assuming equal spacing (e.g., that the
elements of the vector are observations
of a quantity at equally spaced times).
}
\details{
Missing values include \code{NA}, \code{NaN}, \code{Inf}, and \code{-Inf}.
Those not preceded by or not followed by an observed
value cannot be imputed by linear interpolation and
are replaced with the common value \code{NA}.

If \code{zero_as_na} is \code{TRUE}, then zeros preceded by and
followed by an observed value (zero or nonzero) are
treated like \code{NA}. Zeros not preceded by or not followed
by an observed value (zero or nonzero) are left untouched.
}
\examples{
x <- c(NA, 0, NA, 0, NA, 2, NA, 4, 0, 6, NaN, Inf, -Inf)
impute_na(x)
impute_na(x, zero_as_na = TRUE)

}
