% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_beta.R
\name{make_beta}
\alias{make_beta}
\title{\loadmathjax
Define a seasonally forcing function for simulations}
\usage{
make_beta(epsilon, n)
}
\arguments{
\item{epsilon}{A numeric scalar. The standard deviation of
the noise process.}

\item{n}{An integer scalar. The number of observations in
the noise process.}
}
\value{
A function with arguments \code{s} and \code{par_list} (see Details).
}
\description{
Defines a seasonal forcing function with environmental noise
that can be assigned to the argument \code{beta} of \code{\link[=make_data]{make_data()}}.
}
\details{
A noise process \mjseqn{\lbrace(i;X_i)\rbrace_{i=0}^{n-1}}
with \mjseqn{X_i \sim \mathrm{Normal}(0,\epsilon^2)} is
realized and linearly interpolated, yielding a randomly
generated function \mjseqn{a(s)} defined continuously on
the interval \mjseqn{\lbrack 0,n-1 \rbrack}. A seasonal
forcing function \code{f()} is then defined according to

\mjsdeqn{f(s) = \langle\beta\rangle \left\lbrack 1 + \alpha \cos\left(\frac{2 \pi s \Delta t}{\text{365 days}} + a(s)\unicode{x1D7D9}_{\lbrack 0,n-1 \rbrack}\right) \right\rbrack \Delta t}

and returned as output. \code{f()} takes as arguments:

\describe{
\item{\code{s}}{A numeric vector listing values of \mjseqn{s}
at which to evaluate \mjseqn{f(s)}.
}
\item{\code{par_list}}{A list with numeric scalar elements
\code{dt_days}, \code{beta_mean}, and \code{alpha} listing  values for
\mjseqn{\Delta t}, \mjseqn{\langle\beta\rangle}, and
\mjseqn{\alpha}, respectively.
}
}

\code{par_list} does not need an element \code{epsilon} (giving a value
for \mjseqn{\epsilon}), because random number generation and
linear interpolation are performed in the enclosing environment
of \code{f()} (the execution environment of \code{make_beta()}) using the
value of \code{epsilon} found there. That is, \code{f()} does not need to
know \code{epsilon}, because it uses the definition of \mjseqn{a(s)}
that it finds in \code{environment(f)}.

The upshot is that the output of \code{make_beta()} (i.e., the function
\code{f()}) is randomly generated and reproducible with \code{\link[base:Random]{base::set.seed()}},
while the output of \code{f()} is determined and reproducible without
\code{\link[base:Random]{base::set.seed()}}.
}
\examples{
epsilon <- 0.8
n <- 1000
pl <- list(dt_days = 7, beta_mean = 1e-05, alpha = 0.05)
f <- make_beta(epsilon, n)
names(as.list(environment(f)))
a <- get("a", envir = environment(f))
s <- 0:(n-1)
op <- par(mfrow = c(2, 1), mar = c(4, 4, 2, 1))
plot(s, a(s), type = "l")
plot(s, f(s, pl), type = "l")
par(op)

}
\seealso{
\code{\link[=make_data]{make_data()}}
}
