% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_data.R
\name{make_data}
\alias{make_data}
\title{\loadmathjax
Simulate epidemic time series data}
\usage{
make_data(
  par_list = make_par_list(),
  n = 20 * 365/7,
  with_dem_stoch = TRUE,
  ode_control = list(method = "lsoda", rtol = 1e-06, atol = 1e-06)
)
}
\arguments{
\item{par_list}{A list of parameter values containing:
\describe{
\item{\code{dt_weeks}}{\mjeqn{\lbrack \Delta t \rbrack}{[dt]}
Numeric scalar. Observation interval in weeks.
}
\item{\code{t0}}{\mjeqn{\lbrack t_0 \rbrack}{[t_0]}
Numeric scalar. Time of the first observation
in units \mjeqn{\Delta t}{dt}.
}
\item{\code{prep}}{\mjeqn{\lbrack p_\text{rep} \rbrack}{[p_rep]}
Numeric scalar. Case reporting probability.
}
\item{\code{trep}}{\mjeqn{\lbrack t_\text{rep} \rbrack}{[t_rep]}
Numeric scalar. Case reporting delay in units
\mjeqn{\Delta t}{dt}.
}
\item{\code{hatN0}}{\mjeqn{\lbrack \out{\widehat{N}_0} \rbrack}{[hatN_0]}
Numeric scalar. Population size at time \mjseqn{t = 0} years.
}
\item{\code{N0}}{\mjeqn{\lbrack N_0 \rbrack}{[N_0]}
Numeric scalar. Population size at time \mjseqn{t = t_0}.
}
\item{\code{S0}}{\mjeqn{\lbrack S_0 \rbrack}{[S_0]}
Numeric scalar. Number of susceptibles at time \mjseqn{t = t_0}.
}
\item{\code{I0}}{\mjeqn{\lbrack I_0 \rbrack}{[I_0]}
Numeric scalar. Number of infecteds at time \mjseqn{t = t_0}.
}
\item{\code{nu}}{\mjeqn{\lbrack \nu_\text{c} \rbrack}{[nu_c]}
Numeric scalar. Birth rate expressed
per unit \mjeqn{\Delta t}{dt} and
relative to \mjeqn{\out{\widehat{N}_0}}{hatN_0}.
}
\item{\code{mu}}{\mjeqn{\lbrack \mu_\text{c} \rbrack}{[mu_c]}
Numeric scalar. Natural mortality rate expressed
per unit \mjeqn{\Delta t}{dt} and per capita.
}
\item{\code{tgen}}{\mjeqn{\lbrack t_\text{gen} \rbrack}{[t_gen]}
Numeric scalar. Mean generation interval of the disease
of interest in units \mjeqn{\Delta t}{dt}.
}
\item{\code{beta_mean}}{\mjeqn{\lbrack \langle\beta\rangle \rbrack}{[<beta>]}
Numeric scalar. Mean (long-term average) of the seasonally
forced transmission rate \mjeqn{\beta(t)}{beta(t)} expressed
per unit \mjeqn{\Delta t}{dt} per susceptible per infected.
}
\item{\code{alpha}}{\mjeqn{\lbrack \alpha \rbrack}{[alpha]}
Numeric scalar. Amplitude of the seasonally forced transmission
rate \mjeqn{\beta(t)}{beta(t)} relative to the mean.
}
\item{\code{epsilon}}{\mjeqn{\lbrack \epsilon \rbrack}{[epsilon]}
Numeric scalar. Standard deviation of the random phase shift
introduced to the seasonally forced transmission rate
\mjeqn{\beta(t)}{beta(t)}.
}
}}

\item{n}{Integer scalar. Time between the first and last observations
in units \mjeqn{\Delta t}{dt}, so that simulated time series have
\code{n + 1} observations. (If numeric but not integer, then \code{n} is
replaced by \code{floor(n)}.)}

\item{with_dem_stoch}{Logical scalar. If \code{TRUE}, then the simulation
is generated using \code{\link[adaptivetau:ssa.adaptivetau]{adaptivetau::ssa.adaptivetau()}}. Otherwise, it
is generated using \code{\link[deSolve:ode]{deSolve::ode()}} (see Details).}

\item{ode_control}{A list of optional arguments of \code{\link[deSolve:ode]{deSolve::ode()}},
specifying options for numerical integration, such as \code{method},
\code{rtol}, and \code{atol}. Not used if \code{with_dem_stoch = TRUE}.}
}
\value{
A data frame with \code{n + 1} rows corresponding to equally spaced times
\mjeqn{t_k = t_0 + k \Delta t}{t_k = t_0 + k dt}
(for \mjeqn{k = 0, \ldots, n}{k = 0,...,n}), and numeric columns:

\describe{
\item{\code{t}}{Time in units \mjeqn{\Delta t}{dt}. Equal to
\code{par_list$t0 + 0:n}.
}
\item{\code{t_years}}{Time in years. Equal to
\code{(par_list$t0 + 0:n) * par_list$dt_weeks * (7 / 365)}.
}
\item{\code{beta}}{Seasonally forced transmission rate expressed
per unit \mjeqn{\Delta t}{dt} per susceptible per infected,
without environmental noise.
}
\item{\code{beta_phi}}{Seasonally forced transmission rate expressed
per unit \mjeqn{\Delta t}{dt} per susceptible per infected,
with environmental noise.
}
\item{\code{N}}{Population size.}
\item{\code{S}}{Number of susceptibles.}
\item{\code{I}}{Number of infecteds.}
\item{\code{Z}}{Incidence. \code{Z[i]} is the number of infections
between times \code{t[i-1]} and \code{t[i]}.
}
\item{\code{C}}{Reported incidence. \code{C[i]} is the number of infections
reported between times \code{t[i-1]} and \code{t[i]}, equal to the number
of successes in \code{Z[i-round(par_list$trep)]} independent Bernoulli
trials, each with success probability \code{par_list$prep}.
}
\item{\code{B}}{Births. \code{B[i]} is the number of births between times
\code{t[i-1]} and \code{t[i]}.
}
\item{\code{mu}}{Per capita natural mortality rate. \code{mu[i]} is rate
at time \code{t[i]}. Equal to \code{rep(par_list$mu, n + 1)} and included
for convenience, mainly because \code{\link[=estimate_beta_SI]{estimate_beta_SI()}} requires
a data frame with columns \code{C}, \code{B}, and \code{mu}.
}
}

A list of the arguments of \code{make_data()} is included as an attribute.
}
\description{
\code{make_data()} simulates epidemic time series data using a system of
SIR equations and a supplied list of parameter values. Observations
are recorded at equally spaced time points
\mjeqn{t_k = t_0 + k \Delta t}{t_k = t_0 + k dt}
(for \mjeqn{k = 0, \ldots, n}{k = 0,...,n}).
Among other things, the simulation model assumes:
\itemize{
\item A seasonally forced transmission rate.
\item Constant birth and per capita natural mortality rates.
\item Cases reported after a fixed delay since infection,
with a fixed probability.
}
}
\section{Simulation model}{

\code{make_data()} simulates epidemic time series data using the system of
SIR equations below, which includes a fourth equation for cumulative
births and a fifth equation for cumulative incidence:

\mjdeqn{\out{\begin{align} \frac{\text{d}S}{\text{d}t} &= \nu \widehat{N}_0 - \beta_\phi(t) S I - \mu S, \cr \frac{\text{d}I}{\text{d}t} &= \beta_\phi(t) S I - \gamma I - \mu I, \cr \frac{\text{d}R}{\text{d}t} &= \gamma I - \mu R, \cr \frac{\text{d}B_\text{cum}}{\text{d}t} &= \nu \widehat{N}_0, \cr \frac{\text{d}Z_\text{cum}}{\text{d}t} &= \beta_\phi(t) S I. \end{align}}}{(1) dS/dt = nu hatN_0 - beta_phi(t) S I - mu S,    (2) dI/dt = beta_phi(t) S I - gamma I - mu I,    (3) dR/dt = gamma I - mu R,    (4) dB_cum/dt = nu hatN_0,    (5) dZ_cum/dt = beta_phi(t).}

Here, \mjeqn{\gamma = 1 / t_\text{gen}}{gamma = 1 / t_gen},

\mjdeqn{\beta_\phi(t) = \langle\beta\rangle \left\lbrack 1 + \alpha \cos\left(\frac{2 \pi t}{\text{1 year}} + \phi(t;\epsilon)\right) \right\rbrack,}{beta(t) = <beta> (1 + alpha cos(2 pi t / (1 year) + phi(t;epsilon))),}

and \mjeqn{\phi(t;\epsilon)}{phi(t;epsilon)}
is the linear interpolant of noise
\mjeqn{\lbrace(t_k;\Phi_k)\rbrace_{k=0}^n}{{(t_k,Phi_k)}}
with

\mjdeqn{\Phi_k \sim \mathrm{Normal}(0,\epsilon^2),}{Phi_k ~ Normal(0,epsilon^2),}

modeling \strong{environmental stochasticity}.

\code{make_data()} generates observations of the above system at equally
spaced times \mjeqn{t_k = t_0 + k \Delta t}{t_k = t_0 + k dt} (for
\mjeqn{k = 0, \ldots, n}{k = 0,...,n}) by either (i) numerically
integrating the ODE using \code{\link[deSolve:ode]{deSolve::ode()}} (\code{with_dem_stoch = FALSE}),
or (ii) realizing a corresponding continuous-time stochastic process
using \code{\link[adaptivetau:ssa.adaptivetau]{adaptivetau::ssa.adaptivetau()}} (\code{with_dem_stoch = TRUE}).
Both methods use initial state

\mjdeqn{\begin{bmatrix} S(t_0) \cr I(t_0) \cr R(t_0) \cr B_\text{cum}(t_0) \cr Z_\text{cum}(t_0) \end{bmatrix} = \begin{bmatrix} S_0 \cr I_0 \cr N_0 - S_0 - I_0 \cr 0 \cr 0 \end{bmatrix}.}{(S(t_0),I(t_0),R(t_0),B_cum(t_0),Z_cum(t_0)) = (S_0,I_0,N_0 - S_0 - I_0,0,0).}

The latter method defines event probabilities as proportional to
terms in the ODE and models \strong{demographic stochasticity}. Disease
fadeout in simulations with demographic stochasticity is prevented
by setting the probability of infected decrease to zero whenever
the number of infecteds is 1.

\code{make_data()} calculates births \mjseqn{B} and incidence \mjseqn{Z}
from cumulative births \mjeqn{B_\text{cum}}{B_cum} and cumulative
incidence \mjeqn{Z_\text{cum}}{Z_cum} via first differences,

\mjdeqn{\begin{align} B(t_k) &= B_\text{cum}(t_k) - B_\text{cum}(t_{k-1}), \cr Z(t_k) &= Z_\text{cum}(t_k) - Z_\text{cum}(t_{k-1}), \end{align}}{(1) B(t_k) = B_cum(t_k) - B_cum(t_{k-1}),    (2) Z(t_k) = Z_cum(t_k) - Z_cum(t_{k-1}),}

then simulates reported incidence \mjseqn{C} from
incidence \mjseqn{Z} via lagged binomial sampling,

\mjdeqn{C(t_{k+r}) \sim \mathrm{Binomial}\big(Z(t_k),p_\text{rep}\big),}{C(t_{k+r}) ~ Binomial(Z(t_k), p_rep),}

modeling \strong{observation error} (random under-reporting of cases with
a delay between infection and reporting). Here,
\mjeqn{r = \mathrm{nint}(t_\text{rep} / \Delta t)}{r = nint(t_rep / dt)}.
}

\examples{
# Deterministic simulation
par_list <- make_par_list(dt_weeks = 1)
df <- make_data(
  par_list = par_list,
  n = 20 * 365 / 7, # number of weeks in 20 years
  with_dem_stoch = FALSE
)
head(df)

# Stochastic simulation
par_list <- make_par_list(
  dt_weeks = 1,
  epsilon = 0.5, # environmental stochasticity
  prep = 0.5 # random under-reporting of cases
)
set.seed(2146)
df <- make_data(
  par_list = par_list,
  n = 20 * 365 / 7,
  with_dem_stoch = TRUE # demographic stochasticity
)
head(df)

}
