% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_par_list.R
\name{make_par_list}
\alias{make_par_list}
\title{\loadmathjax
Create a list of parameter values for simulations}
\usage{
make_par_list(
  dt_weeks = 1,
  t0 = 1000 * (365/7)/dt_weeks,
  prep = 1,
  trep = 0,
  hatN0 = 1e+06,
  N0 = NULL,
  S0 = NULL,
  I0 = NULL,
  nu = 0.04 * (7/365) * dt_weeks,
  mu = 0.04 * (7/365) * dt_weeks,
  tgen = 13 * (1/7)/dt_weeks,
  Rnaught = 20,
  beta_mean = NULL,
  alpha = 0.08,
  epsilon = 0,
  ode_control = list(method = "lsoda", rtol = 1e-06, atol = 1e-06)
)
}
\arguments{
\item{dt_weeks}{\mjseqn{\lbrack \Delta t \rbrack}
Numeric scalar. Observation interval in weeks.}

\item{t0}{\mjseqn{\lbrack t_0 \rbrack}
Numeric scalar. Time of the first observation
in units \mjseqn{\Delta t}.}

\item{prep}{\mjseqn{\lbrack p_\text{rep} \rbrack}
Numeric scalar. Case reporting probability.}

\item{trep}{\mjseqn{\lbrack m \rbrack}
Numeric scalar. Case reporting delay in units \mjseqn{\Delta t}.}

\item{hatN0}{\mjseqn{\lbrack \widehat{N}_0 \rbrack}
Numeric scalar. Population size at time \mjseqn{t = 0} years.}

\item{N0}{\mjseqn{\lbrack N_0 \rbrack}
Numeric scalar. Population size at time \mjseqn{t = t_0}.
This argument is optional (see Details 2).}

\item{S0}{\mjseqn{\lbrack S_0 \rbrack}
Numeric scalar. Number of susceptibles at time \mjseqn{t = t_0}.
This argument is optional (see Details 2).}

\item{I0}{\mjseqn{\lbrack I_0 \rbrack}
Numeric scalar. Number of infecteds at time \mjseqn{t = t_0}.
This argument is optional (see Details 2).}

\item{nu}{\mjseqn{\lbrack \nu \rbrack}
Numeric scalar. Birth rate expressed per unit \mjseqn{\Delta t}
and relative to \mjseqn{\widehat{N}_0}.}

\item{mu}{\mjseqn{\lbrack \mu \rbrack}
Numeric scalar. Natural mortality rate expressed per unit
\mjseqn{\Delta t} and per capita.}

\item{tgen}{\mjseqn{\lbrack t_\text{gen} \rbrack}
Numeric scalar. Mean generation interval of the disease
of interest in units \mjseqn{\Delta t}.}

\item{Rnaught}{\mjseqn{\lbrack \mathcal{R}_0 \rbrack}
Numeric scalar. Basic reproduction number of the disease
of interest. If specified (not \code{NULL}), then \code{beta_mean} is
calculated internally as a function of \code{Rnaught} (see Details 1).}

\item{beta_mean}{\mjseqn{\lbrack \langle\beta\rangle \rbrack}
Numeric scalar. Mean (long-term average) of the seasonally
forced transmission rate \mjseqn{\beta(t)} expressed per unit
\mjseqn{\Delta t} per susceptible per infected.
Ignored if \code{Rnaught} is specified (not \code{NULL}) (see Details 1).}

\item{alpha}{\mjseqn{\lbrack \alpha \rbrack}
Numeric scalar. Amplitude of the seasonally forced transmission
rate \mjseqn{\beta(t)} relative to the mean.}

\item{epsilon}{\mjseqn{\lbrack \epsilon \rbrack}
Numeric scalar. Standard deviation of the random phase shift
introduced to the seasonally forced transmission rate
\mjseqn{\beta(t)}.}

\item{ode_control}{A list of optional arguments of \code{\link[deSolve:ode]{deSolve::ode()}},
specifying options for numerical integration, such as \code{method},
\code{rtol}, and \code{atol} (see Details 2).}
}
\value{
A list of the arguments of \code{make_par_list()} (excluding
\code{ode_control}) with values for \code{Rnaught}, \code{beta_mean},
\code{N0}, \code{S0}, and \code{I0} if not supplied in the function call
(see Details 1 and 2).
}
\description{
\code{make_par_list()} creates a list of parameter values to be passed as
an argument of \code{\link[=make_data]{make_data()}}, which simulates epidemic time series data
using an SIR model.
}
\section{Details}{
\subsection{1. Concordance of \mjseqn{\mathcal{R}_0} and \mjseqn{\langle\beta\rangle}}{

\code{make_par_list()} enforces the identity

\mjsdeqn{\mathcal{R}_0 = \frac{\nu \widehat{N}_0}{\mu} \cdot \frac{\langle\beta\rangle}{\gamma + \mu}\,,}

where \mjseqn{\gamma = 1 / t_\text{gen}}. This means that it calculates
\mjseqn{\langle\beta\rangle} as a function of \mjseqn{\mathcal{R}_0}, unless
the user explicitly sets \code{Rnaught = NULL}, in which case it calculates
\mjseqn{\mathcal{R}_0} as a function of \mjseqn{\langle\beta\rangle}. Hence
argument \code{beta_mean} is optional if \code{Rnaught} is specified, but mandatory
otherwise.
}

\subsection{2. Missing \mjseqn{N_0}, \mjseqn{S_0}, or \mjseqn{I_0}}{

If \mjseqn{N_0}, \mjseqn{S_0}, or \mjseqn{I_0} (\code{N0}, \code{S0}, or \code{I0})
is not specified in the function call, then, with a call to
\code{\link[deSolve:ode]{deSolve::ode()}}, \code{make_par_list()} numerically integrates the
system of SIR equations

\mjsdeqn{\begin{align} \frac{\text{d}S}{\text{d}t} &= \nu \widehat{N}_0 - \beta(t) S I - \mu S \cr \frac{\text{d}I}{\text{d}t} &= \beta(t) S I - \gamma I - \mu I \cr \frac{\text{d}R}{\text{d}t} &= \gamma I - \mu R \end{align}}

with \mjseqn{\gamma = 1 / t_\text{gen}} and

\mjsdeqn{\beta(t) = \langle\beta\rangle \left\lbrack 1 + \alpha \cos\left(\frac{2 \pi t}{\text{1 year}}\right) \right\rbrack}

between times \mjseqn{t = 0} years and \mjseqn{t = t_0}, taking

\mjsdeqn{\begin{bmatrix} S(0) \cr I(0) \cr R(0) \end{bmatrix} = \widehat{N}_0 \begin{bmatrix} \frac{1}{\mathcal{R}_0} \cr \big(1 - \frac{1}{\mathcal{R}_0}\big) \frac{\mu}{\gamma + \mu} \cr \big(1 - \frac{1}{\mathcal{R}_0}\big) \frac{\gamma}{\gamma + \mu} \end{bmatrix}}

for the initial state. This is the endemic equilibrium of the above system
with constant transmission rate \mjseqn{\beta(t) \equiv \langle\beta\rangle},
scaled by \mjseqn{\mu/\nu} to enforce
\mjseqn{S(0) + I(0) + R(0) = \widehat{N}_0}

as the initial population size. Then \code{make_par_list()} assigns \code{N0},
\code{S0}, and \code{I0} (only those not specified in the function call) the
values

\describe{
\item{\code{N0}}{\mjseqn{S(t_0) + I(t_0) + R(t_0)}}
\item{\code{S0}}{\mjseqn{S(t_0)}}
\item{\code{I0}}{\mjseqn{I(t_0)}}
}

A warning is issued if the ODE solver cannot complete
the integration. A different solver may have more success
(e.g., consider \code{ode_control = list(method = "ode45")}
to override the default \code{method = "lsoda"}).
Using different \code{rtol} and \code{atol} might also help.
}
}

\examples{
# Creates a reasonable list without user input
par_list <- make_par_list()
unlist(par_list)

# Requires time and rate parameters
# in units of the observation interval
dt_weeks <- 1
nu_peryear <- 0.04
tgen_days <- 13
par_list <- make_par_list(
  dt_weeks = dt_weeks,
  nu       = nu_peryear * (7 / 365) * dt_weeks,
  tgen     = tgen_days * (1 / 7) / dt_weeks
)
unlist(par_list)

}
