% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peaks.R
\name{peaks-methods}
\alias{peaks-methods}
\alias{plot.peaks}
\alias{print.peaks}
\title{Methods for class peaks}
\usage{
\method{plot}{peaks}(x, ...)

\method{print}{peaks}(x, ...)
}
\arguments{
\item{x}{A peaks object.}

\item{...}{Unused optional arguments.}
}
\description{
Methods for plotting and printing peaks objects
returned by \code{\link[=peaks]{peaks()}}.
}
