% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peaks.R
\name{peaks}
\alias{peaks}
\title{\loadmathjax
Locate peaks in time series}
\usage{
peaks(x, bw1, bw2, period = NULL)
}
\arguments{
\item{x}{Numeric vector. An equally spaced time series.
Defines \mjseqn{\lbrace x_i \rbrace_{i=0}^{n-1}} in Algorithm.
(Without loss of generality, the indices \code{seq_along(x)} are
used as time points, so that \mjseqn{t_i = i+1} for all \mjseqn{i}.)}

\item{bw1}{Integer scalar. Bandwidth for the central moving average
applied to \code{x}, so that \code{xbar[i] = mean(x[(i-bw1):(i+bw1)])} for
all \code{i}. Defines \mjseqn{\ell_1} in Algorithm.}

\item{bw2}{Integer scalar. Bandwidth for peak definition, so that a
peak occurs at index \code{i} if and only if
\code{all(xbar[i] > xbar[(i-bw2):(i-1)])} and
\code{all(xbar[i] > xbar[(i+1):(i+bw2)])}.
Defines \mjseqn{\ell_2} in Algorithm.}

\item{period}{Numeric scalar. Period of \code{x} if \code{x} is roughly periodic.
Defines \mjseqn{T} in Algorithm. This argument is optional and should
retain the default value \code{NULL} if \code{x} is not roughly periodic.}
}
\value{
A peaks object. A list with elements:

\describe{
\item{\code{xbar}}{Numeric vector. The result of applying
a \code{(2 * bw1 + 1)}-point central moving average to \code{x}.
}
\item{\code{all}}{Numeric vector. A subset of \code{seq_along(x)}
listing the index of each peak in \code{x}.
}
\item{\code{phase}}{Numeric vector. A subset of \code{all} listing
the index of each peak in phase with the first peak.
\code{NULL} if \code{period} is \code{NULL} in the function call.
}
}

The list has attributes \code{call} and \code{arg_list}, making it
reproducible with \code{eval(call)} or \code{do.call(peaks, arg_list)}.
}
\description{
Locates peaks in time series with observations at equally spaced
time points \mjseqn{t_i = t_0 + i \Delta t}. For roughly periodic
time series, identifies those peaks in phase with the first peak.
}
\section{Details}{
\subsection{1. Algorithm}{

The aim is to index peaks in the supplied time series
\mjseqn{\lbrace (t_i,x_i) \rbrace_{i=0}^{n-1}}. To ensure that
this output is robust to noise, a \mjseqn{(2 \ell_1 + 1)}-point
central moving average is applied, generating a smoother time series
\mjseqn{\out{\lbrace (t_i,\bar{x}_i) \rbrace_{i=\ell_1}^{n-1-\ell_1}}}
with

\mjsdeqn{\out{\bar{x}_i = \frac{1}{2 \ell_1 + 1} \sum_{j=-\ell_1}^{\ell_1} x_{i+j}\,.}}

Peaks in
\mjseqn{\out{\lbrace (t_i,\bar{x}_i) \rbrace_{i=\ell_1}^{n-1-\ell_1}}}
are found by taking temporal neighbourhoods containing
\mjseqn{(2 \ell_2 + 1)} points and checking whether the central point
is a local maximum. Specifically, \mjseqn{(t_k,\bar{x}_k)} is defined
as a peak if and only if \mjseqn{k \in \mathcal{I}}, where

\mjsdeqn{\out{\mathcal{I} = \Big\lbrace i \in \lbrace\ell_1+\ell_2,\ldots,n-1-\ell_1-\ell_2\rbrace : \bar{x}_i > \bar{x}_{i \pm j}\,\text{for all}\,j = 1,\ldots,\ell_2 \Big\rbrace\,.}}

If the unobserved function \mjseqn{x(t)} is \mjseqn{T}-periodic,
then a secondary aim is to identify all peaks in phase with the
first peak at \mjseqn{\out{(t_a,\bar{x}_a)}},
where \mjseqn{a = \min(\mathcal{I})}.
\mjseqn{\out{(t_k,\bar{x}_k)}} is defined as a peak in phase with
the first peak if and only if
\mjseqn{k \in \mathcal{I}_\text{phase} \subset \mathcal{I}}, where

\mjsdeqn{\out{\mathcal{I}_\text{phase} = \Big\lbrace \arg\min_{i \in \mathcal{I}} \big|t_i - (t_a + jT)\big| : j = 0,\ldots,m \Big\rbrace\,.}}

Here, \mjseqn{m} is the number of complete cycles between times
\mjseqn{t_a} and \mjseqn{t_{n-1-\ell_1}}, given by

\mjsdeqn{m = \left\lfloor \frac{t_{n-1-\ell_1} - t_a}{T} \right\rfloor\,.}

Hence a peak is considered in phase with the first peak if and
only if it occurs nearest to an integer multiple of the period
after the first peak.
}

\subsection{2. Missed peaks}{

The procedure outlined in Algorithm will not detect peaks at the edges
of \mjseqn{\lbrace (t_i,x_i) \rbrace_{i=0}^{n-1}}. Specifically, it will
not detect a peak at time \mjseqn{t_k} if \mjseqn{k < \ell_1+\ell_2} or
\mjseqn{k > n-1-\ell_1-\ell_2}. It can miss peaks in other (unlucky)
situations. For example, a peak at time \mjseqn{t_k} will be missed if,
by chance, \mjseqn{\out{\bar{x}_k = \bar{x}_{k+1}}} (i.e., if, after
smoothing, there are two points of equal height at the top of a peak
instead of a single highest point).
}

\subsection{3. Bandwidth tuning}{

The bandwidths \code{bw1} and \code{bw1}(denoted by \mjseqn{\ell_1} and
\mjseqn{\ell_2} in Algorithm) must be tuned in order to disqualify
spurious peaks in \mjseqn{\lbrace (t_i,x_i) \rbrace_{i=0}^{n-1}}
caused by noise, without disqualifying true peaks. Disqualifying
spurious peaks is a matter of smoothing to a sufficient degree
(choosing \code{bw1} large enough) and including a nontrivial number
of points in the temporal neighbourhoods that distinguish local
maxima from other points (choosing \code{bw2} large enough). Not
disqualifying true peaks is a matter of not including more than
one peak in those temporal neighbourhoods (choosing \code{bw2} less
than \code{0.5 * num_obs_in_period / num_peaks_in_period}).

Reasonable results are typically obtained by:
\itemize{
\item Choosing the smallest \code{bw1} such that
\mjseqn{\out{\lbrace (t_k,\bar{x}_k) \rbrace_{k=\ell_1}^{n-1-\ell_1}}}
is smooth around the peaks.
\item Choosing \code{bw2} around
\code{0.4 * num_obs_in_period / num_peaks_in_period}.
A good value for \code{bw1} can be found by experimenting with \code{\link[stats:filter]{stats::filter()}}
(see Examples).
}
}
}

\examples{
# Create a noisy periodic time series
times <- seq(0, 20, by = 0.01)
x <- sin(2 * pi * times) + sin(pi * times) + rnorm(times, 0, 0.5)
plot(x, type = "l", lwd = 2, col = "grey80", las = 1)

# Smoothing with `bw1 = 20` removes noise around the peaks
bw1 <- 20
m <- 2 * bw1 + 1
xbar <- as.vector(
  filter(x,
    filter = rep(1 / m, m),
    method = "convolution",
    sides  = 2
  )
)
lines(xbar, lwd = 2)

# Setting `bw2 = 0.4 * num_obs_in_period / num_peaks_in_period`
# tends to yield reasonable results
bw2 <- 0.4 * 200 / 2

# Locate peaks in time series
peaks_out <- peaks(
  x = x,
  bw1 = bw1,
  bw2 = bw2,
  period = 200 # number of observations in a period (2 units of time)
)

# Verify that peaks were identified correctly:
# * Blue lines are peaks.
# * Pink circles are peaks in phase with first peak.
plot(peaks_out)

}
\references{
\insertRef{Jaga+20}{fastbeta}
}
\seealso{
\code{\link[=ptpi]{ptpi()}}, \code{\link[=plot.peaks]{plot.peaks()}}
}
