% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.fastbeta.R
\name{plot.fastbeta}
\alias{plot.fastbeta}
\title{Plot a fastbeta object}
\usage{
\method{plot}{fastbeta}(x, ...)
}
\arguments{
\item{x}{A fastbeta object.}

\item{...}{Unused optional arguments.}
}
\value{
\code{NULL} is returned invisibly.
}
\description{
Plots a fastbeta object.
}
