% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptpi.R
\name{ptpi-methods}
\alias{ptpi-methods}
\alias{plot.ptpi}
\alias{print.ptpi}
\title{Methods for class "ptpi"}
\usage{
\method{plot}{ptpi}(x, ...)

\method{print}{ptpi}(x, ...)
}
\arguments{
\item{x}{A ptpi object.}

\item{...}{Unused optional arguments.}
}
\description{
Methods for plotting and printing ptpi objects
returned by \code{\link[=ptpi]{ptpi()}}.
}
