% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ptpi.R
\name{ptpi}
\alias{ptpi}
\title{\loadmathjax
Estimate \mjseqn{S_0} using PTPI}
\usage{
ptpi(data, S0_init, peak1 = 1L, peak2 = nrow(data), it = 10L)
}
\arguments{
\item{data}{A data frame with numeric columns:

\describe{
\item{\code{t}}{\mjseqn{\lbrace\,t_i\,\rbrace}
Time. Lists the observation times \mjseqn{t_i = t_0 + i \Delta t}
(for \mjseqn{i = 0, \ldots, n-1}) in any convenient units. Here,
\mjseqn{t_0} is the initial observation time and \mjseqn{\Delta t}
is the observation interval.
}
\item{\code{Z}}{\mjseqn{\lbrace\,Z_i\,\rbrace}
Incidence. \code{Z[i]} is the number of infections
between times \code{t[i-1]} and \code{t[i]}. Must be roughly periodic.
}
\item{\code{B}}{\mjseqn{\lbrace\,B_i\,\rbrace}
Births. \code{B[i]} is the number of births
between times \code{t[i-1]} and \code{t[i]}.
}
\item{\code{mu}}{\mjseqn{\lbrace\,\mu_i \Delta t\,\rbrace}
Per capita natural mortality rate. \code{mu[i]} is the rate
at time \code{t[i]} expressed per unit \mjseqn{\Delta t}.
}
}}

\item{S0_init}{A numeric scalar. An initial estimate of \mjseqn{S_0}.}

\item{peak1}{An integer scalar. Index of the first peak in \code{data$Z}.
A reasonable value can be obtained using \code{\link[=peaks]{peaks()}} (see Examples).}

\item{peak2}{An integer scalar. Index of the last peak in \code{data$Z}
in phase with the first peak.
A reasonable value can be obtained using \code{\link[=peaks]{peaks()}} (see Examples).}

\item{it}{An integer scalar. The number of iterations to perform
before stopping.}
}
\value{
A ptpi object. A list with elements:

\describe{
\item{\code{times}}{A numeric vector listing times since the initial
time in units of the observation interval. The rows of \code{mat}
correspond to these time points. Equal to \code{0:(nrow(data)-1)}.
}
\item{\code{mat}}{A numeric matrix with \code{length(times)} rows and \code{1+it}
columns containing the susceptible time series generated in each
iteration. \code{mat[i,j]} gives the value of \mjseqn{S_{i-1}^{(j-1)}}
as defined in Algorithm.
}
\item{\code{S0}}{A numeric vector listing in order all \code{1+it} estimates
of \mjseqn{S_0 = S(t_0)}. Equivalent to \code{mat[1, ]}.
}
\item{\code{S0_final}}{A numeric scalar giving the final estimate
of \mjseqn{S_0}. Equivalent to \code{mat[1, ncol(mat)]}.
}
\item{\code{call}}{The function call. The ptpi object is reproducible
with \code{eval(call)}.
}
\item{\code{arg_list}}{A list of the arguments in the function call.
The ptpi object is reproducible with \code{do.call(ptpi, arg_list)}.
}
}
}
\description{
Using peak-to-peak iteration (PTPI, see Algorithm), estimates the
initial number of susceptible individuals \mjseqn{S_0 = S(t_0)}
from time series of incidence, births, and natural mortality with
observations at equally spaced time points \mjseqn{t_i = t_0 + i \Delta t}.
The incidence time series must be roughly periodic.
}
\section{Details}{
\subsection{1. Algorithm}{

Supplied in the data frame \code{data} are time series
\mjseqn{\lbrace (t_i,Z_i) \rbrace_{i=0}^{n-1}},
\mjseqn{\lbrace (t_i,B_i) \rbrace_{i=0}^{n-1}}, and
\mjseqn{\lbrace (t_i,\mu_i) \rbrace_{i=0}^{n-1}}
of incidence, births, and natural mortality with observations
at equally spaced time points \mjseqn{t_i = t_0 + i \Delta t}.
The first peak in incidence and the last peak in phase with
the first peak occur at times \mjseqn{t_a} and \mjseqn{t_b},
respectively, where \mjseqn{a} is \code{peak1-1} and \mjseqn{b}
is \code{peak2-1}.

Let \mjseqn{S_i^{(j)}} denote the estimate of \mjseqn{S(t_i)}
obtained after \mjseqn{j} iterations. Assign both \mjseqn{S_0^{(0)}}
and \mjseqn{S_a^{(0)}} the value of \code{S0_init}, then proceed with
the following iteration.

Given \mjseqn{S_a^{(j)}} for some \mjseqn{j}, define \mjseqn{S_i^{(j)}}
for all \mjseqn{i \in \lbrace a+1,\ldots,n-1 \rbrace} with the recursion

\mjsdeqn{S_{i+1}^{(j)} = \frac{\big\lbrack 1 - \frac{1}{2} \mu_i \Delta t \big\rbrack S_i^{(j)} + B_{i+1} - Z_{i+1}}{1 + \frac{1}{2} \mu_{i+1} \Delta t}\,.}

Then set \mjseqn{S_a^{(j+1)} = S_b^{(j)}} and define \mjseqn{S_i^{(j+1)}}
for all \mjseqn{i \in \lbrace 0,\ldots,a-1 \rbrace} with the recursion

\mjsdeqn{S_{i-1}^{(j+1)} = \frac{\big\lbrack 1 + \frac{1}{2} \mu_i \Delta t \big\rbrack S_i^{(j+1)} - B_i + Z_i}{1 + \frac{1}{2} \mu_{i-1} \Delta t}\,.}

Note that this second recursion is carried out backwards in time.
}

\subsection{2. Convergence}{

Suppose that incidence is \mjseqn{(p \Delta t)}-periodic and
(without loss of generality) that \mjseqn{b - a = m p}, where
\mjseqn{p} and \mjseqn{m} are positive integers. Then
\mjseqn{Z_i = Z_{i+kp}} for all integers \mjseqn{i} and \mjseqn{k},
and there are \mjseqn{m} periods between times \mjseqn{t_a} and
\mjseqn{t_b}. Now consider the special case in which the birth
and per capita death rates are constant, so that \mjseqn{B_i = B}
and \mjseqn{\mu_i = \mu} for all integers \mjseqn{i}.
In this case, one can show that the sequence \mjseqn{(S_0^{(j)})}
converges to a limit \mjseqn{S_0^*} given by

\mjsdeqn{S_0^* = \left( \frac{1 + \frac{1}{2} \mu \Delta t}{1 - \frac{1}{2} \mu \Delta t} \right)^a S_a^* + \frac{1}{1 - \frac{1}{2} \mu \Delta t} \sum_{i=1}^a (Z_i - B) \left( \frac{1 + \frac{1}{2} \mu \Delta t}{1 - \frac{1}{2} \mu \Delta t} \right)^{i-1}\,,}

where \mjseqn{S_a^*} is the limit of the sequence \mjseqn{(S_a^{(j)})},
given by

\mjsdeqn{S_a^* = \frac{B}{\mu \Delta t} - \frac{1}{1 + \frac{1}{2} \mu \Delta t} \left\lbrack \frac{\left( \frac{1 - \frac{1}{2} \mu \Delta t}{1 + \frac{1}{2} \mu \Delta t} \right)^p}{1 - \left( \frac{1 - \frac{1}{2} \mu \Delta t}{1 + \frac{1}{2} \mu \Delta t} \right)^p} \right\rbrack \sum_{i=1}^p Z_{a+i} \left( \frac{1 + \frac{1}{2} \mu \Delta t}{1 - \frac{1}{2} \mu \Delta t} \right)^i\,.}

One can further show that the convergence is linear and follows

\mjsdeqn{S_0^{(j)} - S_0^* = (S_0^{(0)} - S_a^*) \left( \frac{1 - \frac{1}{2} \mu \Delta t}{1 + \frac{1}{2} \mu \Delta t} \right)^{jp-a}\,.}

A corollary is that convergence of \mjseqn{(S_0^{(j)})} is guaranteed
even if incidence is not truly periodic, because, within the iteration,
incidence is always \emph{effectively} \mjseqn{((b - a) \Delta t)}-periodic.
This can be seen by noting that \mjseqn{S_a^{(j)}} is updated recursively,
with each update using the same \mjseqn{b - a} observations of incidence:
\mjseqn{Z_{a+1}, \ldots, Z_b}. Hence the limits above will be correct
even for aperiodic (or just roughly periodic) incidence provided that
\mjseqn{p} is replaced with \mjseqn{b - a}.
}
}

\examples{
## Simulate 20 years of disease incidence,
## observed weekly
pl <- make_par_list(model = "sir")
df <- make_data(pl,
  n = 20 * 365 / 7, # number of weeks in 20 years
  with_ds = TRUE,
  model = "sir"
)

## Plot incidence time series, and note the
## apparent 1-year period
plot(Z ~ t_years, df, type = "l",
  xlab = "Time (years)",
  ylab = "Incidence"
)

## Locate peaks in incidence time series
peaks_out <- peaks(
  x = df$Z,
  bw1 = 6,
  bw2 = 8,
  period = 365 / 7 # number of weeks in 1 year
)

## Verify that peaks were identified correctly:
## * Blue lines are peaks.
## * Pink circles are peaks in phase with first peak.
plot(peaks_out)

## Estimate `S0` from incidence, births, and
## natural mortality using PTPI, starting from
## an erroneous initial guess
ptpi_out <- ptpi(
  data = df,
  S0_init = df$S[1] * 4,
  peak1 = with(peaks_out, phase[1]),
  peak2 = with(peaks_out, phase[length(phase)]),
  it = 25
)
plot(ptpi_out)

## Sequence of estimates
ptpi_out$S0

## Relative error in final estimate
(ptpi_out$S0_final - df$S[1]) / df$S[1]

}
\references{
\insertRef{Jaga+20}{fastbeta}
}
\seealso{
\link[=ptpi-methods]{methods for class "ptpi"}, \code{\link[=peaks]{peaks()}}
}
