% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/try_loess.R
\name{try_loess}
\alias{try_loess}
\title{Explore loess fits to equally spaced time series}
\usage{
try_loess(formula, data, q = 6:9, ..., plot = TRUE)
}
\arguments{
\item{formula}{A formula specifying a time series in \code{data}.
Something like \code{x ~ t}.}

\item{data}{A data frame containing time series data.
Must satisfy \code{all(all.vars(formula) \%in\% names(data))}
(column names match variable names),
\code{length(unique(diff(data[[all.vars(formula)[2]]]))) == 1}
(time column is equally spaced).}

\item{q}{An integer vector. The values of \code{span} passed to
\code{\link[stats:loess]{stats::loess()}} will be \code{q / nrow(data)} (see Details).}

\item{...}{Additional arguments of \code{\link[stats:loess]{stats::loess()}}, such as
\code{degree}, but not \code{subset} or \code{span}. Those not specified
are assigned their default values (see \code{\link[stats:loess]{stats::loess()}}).}

\item{plot}{A logical scalar. If \code{TRUE}, then the fitted curves
are plotted (4 curves per plot).}
}
\value{
Invisibly, a list of loess objects with names \code{paste0("q", q)}.
}
\description{
A wrapper of \code{\link[stats:loess]{stats::loess()}} for equally spaced time series.
Fits loess curves to the specified time series using each
supplied value of the smoothing parameter and optionally plots
the fitted curves.
}
\details{
When applying \code{\link[stats:loess]{stats::loess()}} to smooth an equally spaced
time series, all possible smoothing kernels are captured by
\code{span = q / nrow(data)} with integers \code{q}. All other values
of \code{span} (those produced by non-integer \code{q}) are redundant.
See \insertCite{ClevGros+91;textual}{fastbeta} for details.

When fitting local quadratic polynomials (\code{degree = 2};
the default), \code{q > 5} is required to ensure that the local
smoothing kernel assigns at least 3 observations a positive
weight. This will guarantee a unique least squares fit.
When fitting local linear polynomials (\code{degree = 1}),
\code{q > 3} is sufficient.

To avoid k-d tree warnings, set the optional argument
\code{control = loess.control(surface = "direct")}.
See \code{\link[stats:loess.control]{stats::loess.control()}} and
\insertCite{ClevGros+91;textual}{fastbeta} for details.
}
\examples{
times <- seq(0, 4, by = 0.01)
x <- sin(2 * pi * times) + rnorm(times, 0, 1)
df <- data.frame(t = times, x)
q <- c(6, 20, 100, 400)
control <- loess.control(surface = "direct")
try_loess_out <- try_loess(x ~ t, df, q, control = control, plot = TRUE)

## Fit with `q = 100` looks okay
my_loess <- try_loess_out[["q100"]]

}
\references{
\insertRef{ClevGros+91}{fastbeta}
}
