% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_rrmse.R
\name{compute_rrmse}
\alias{compute_rrmse}
\title{Compute RRMSE in a time series estimate}
\usage{
compute_rrmse(act, est, na_rm = TRUE)
}
\arguments{
\item{act}{Numeric vector. A time series.}

\item{est}{Numeric vector with length equal to \code{length(act)}.
A time series estimate of \code{act}.}

\item{na_rm}{Logical. If \code{TRUE}, then missing values
(\code{NA}, \code{NaN}, \code{Inf}) are ignored.}
}
\value{
Let \code{cc} ("complete cases") be the vector of indices \code{i} such that
neither \code{act[i]} nor \code{est[i]} is missing. If \code{length(cc) = 0} or if
\code{length(cc) < length(act)} and \code{na_rm = FALSE}, then \code{NA}. Otherwise,
the RRMSE in \code{est} given by
\code{sqrt(sum(act[cc] - est[cc])^2 / length(cc)) / mean(act[cc])}.
}
\description{
\code{compute_rrmse()} computes the relative root mean square error
(RRMSE) in a time series estimate of a known time-varying quantity.
}
