% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate-beta.R
\name{estimate-beta}
\alias{estimate-beta}
\alias{estimate_beta_fc}
\alias{estimate_beta_s}
\alias{estimate_beta_si}
\alias{estimate_beta_sei}
\title{\loadmathjax
Estimate time-varying transmission rates}
\usage{
estimate_beta_fc(data, par_list)

estimate_beta_s(data, par_list)

estimate_beta_si(data, par_list)

estimate_beta_sei(data, par_list)
}
\arguments{
\item{data}{A data frame with numeric columns:

\describe{
\item{\code{t}}{\mjseqn{\lbrace\,t_i\,\rbrace}
Time. Lists the observation times \mjseqn{t_i = t_0 + i \Delta t}
(for \mjseqn{i = 0, \ldots, n-1}) in any convenient units. Here,
\mjseqn{t_0} is the initial observation time and \mjseqn{\Delta t}
is the observation interval.
}
\item{\code{Z}}{\mjseqn{\lbrace\,Z_i\,\rbrace}
Incidence. \code{Z[i]} is the number of infections
between times \code{t[i-1]} and \code{t[i]}.
}
\item{\code{B}}{\mjseqn{\lbrace\,B_i\,\rbrace}
Births. \code{B[i]} is the number of births
between times \code{t[i-1]} and \code{t[i]}.
}
\item{\code{mu}}{\mjseqn{\lbrace\,\mu_i \Delta t\,\rbrace}
Per capita natural mortality rate. \code{mu[i]} is the rate
at time \code{t[i]} expressed per unit \mjseqn{\Delta t}.
S, SI, and SEI methods only.
}
}}

\item{par_list}{A list with numeric scalar elements:

\describe{
\item{\code{tgen}}{\mjseqn{\lbrace\,t_\text{gen} / \Delta t\,\rbrace}
Mean generation interval of the disease of interest
in units \mjseqn{\Delta t}. FC, S, and SI methods only.
}
\item{\code{tlat}}{\mjseqn{\lbrace\,t_\text{lat} / \Delta t\,\rbrace}
Mean latent period of the disease of interest
in units \mjseqn{\Delta t}. SEI method only.
}
\item{\code{tinf}}{\mjseqn{\lbrace\,t_\text{inf} / \Delta t\,\rbrace}
Mean infectious period of the disease of interest
in units \mjseqn{\Delta t}. SEI method only.
}
\item{\code{S0}}{\mjseqn{\lbrace\,S_0\,\rbrace}
Number of susceptible individuals at time \mjseqn{t = t_0}.
}
\item{\code{E0}}{\mjseqn{\lbrace\,E_0\,\rbrace}
Number of exposed (infected but not infectious) individuals
at time \mjseqn{t = t_0}. SEI method only.
}
\item{\code{I0}}{\mjseqn{\lbrace\,I_0\,\rbrace}
Number of infectious individuals at time \mjseqn{t = t_0}.
SI and SEI methods only.
}
}}
}
\value{
A data frame with numeric columns:

\describe{
\item{\code{t}}{\mjseqn{\lbrace\,t_i\,\rbrace}
Time. Identical to \code{data$t}.
}
\item{\code{Z}}{\mjseqn{\lbrace\,Z_i\,\rbrace}
Incidence. Identical to \code{data$Z}.
}
\item{\code{Z_agg}}{\mjseqn{\lbrace\,Z_i^\text{agg}\,\rbrace}
Aggregated incidence. \code{Z_agg[i]} is the number of infections
between times \code{t[i-round(tgen)+1]} and \code{t[i]}, for \code{i} in
\code{seq(1 + round(tgen), nrow(data), by = round(tgen))}.
\code{Z_agg[i]} is \code{NA} for all other \code{i}. FC method only.
}
\item{\code{B}}{\mjseqn{\lbrace\,B_i\,\rbrace}
Births. Identical to \code{data$B}.
}
\item{\code{B_agg}}{\mjseqn{\lbrace\,B_i^\text{agg}\,\rbrace}
Aggregated births. \code{B_agg[i]} is the number of births
between times \code{t[i-round(tgen)+1]} and \code{t[i]}, for \code{i} in
\code{seq(1 + round(tgen), nrow(data), by = round(tgen))}.
\code{B_agg[i]} is \code{NA} for all other \code{i}. FC method only.
}
\item{\code{mu}}{\mjseqn{\lbrace\,\mu_i \Delta t\,\rbrace}
Per capita natural mortality rate. Identical to \code{data$mu}.
S, SI, and SEI methods only.
}
\item{\code{S}}{\mjseqn{\lbrace\,S_i\,\rbrace}
Estimated number of susceptible individuals.
}
\item{\code{E}}{\mjseqn{\lbrace\,E_i\,\rbrace}
Estimated number of exposed (infected but not infectious) individuals.
SEI method only.
}
\item{\code{I}}{\mjseqn{\lbrace\,I_i\,\rbrace}
Estimated number of infectious individuals.
}
\item{\code{beta}}{\mjseqn{\lbrace\,\beta_i \Delta t\,\rbrace}
Estimated transmission rate, expressed per unit \mjseqn{\Delta t}
per susceptible individual per infectious individual.
}
}
}
\description{
Functions implementing the FC, S, SI, and SEI methods (see Algorithm)
for estimating time-varying transmission rates \mjseqn{\beta(t)} from
time series of incidence, births, and natural mortality with observations
at equally spaced time points \mjseqn{t_i = t_0 + i \Delta t}. The FC
and S methods are deprecated and outperformed by the more robust SI
and SEI methods. These four functions should not be called directly,
as they are all accessible through the \code{method} argument of the more
useful constructor function \code{\link[=fastbeta]{fastbeta()}}, which does checks on input
and output and returns a fastbeta object with associated methods.
}
\section{Details}{
\subsection{1. Algorithm}{
\subsection{FC method}{

The supplied incidence time series \mjseqn{Z_i} (\code{data$Z}) is aggregated
over generation intervals:

\mjsdeqn{Z_i^\text{agg} = \sum_{k=i-g+1}^k Z_k\,,\quad i=jg\,,\quad j = 1,2,\ldots\,,}

where \mjseqn{g = \mathrm{nint}(t_\text{gen} / \Delta t)} is the mean
generation interval \mjseqn{t_\text{gen}} in units of the observation
interval \mjseqn{\Delta t} (\code{par_list$tgen}), rounded to the nearest
integer. The supplied births time series \mjseqn{B_i} (\code{data$B})
is aggregated similarly. The susceptible population size is estimated
recursively starting from the supplied initial value \mjseqn{S_0}
(\code{par_list$S0}):

\mjsdeqn{S_{i+g} = S_i + B_{i+g}^\text{agg} - Z_{i+g}^\text{agg}\,,\quad i=jg\,,\quad j = 0,1,\ldots\,,}

where natural mortality of susceptibles is assumed to be negligible.
The infectious population size is approximated by aggregated incidence:

\mjsdeqn{I_i = Z_i^\text{agg}\,,\quad i=jg\,,\quad j = 1,2,\ldots\,,}

where natural mortality of infectious individuals is assumed to
be negligible. The transmission rate per susceptible individual
per infectious individual is then estimated as

\mjsdeqn{\beta_i = \frac{Z_{i+g}}{S_i I_i g \Delta t}\,,\quad i=jg\,,\quad j = 1,2,\ldots\,.}

Note that this algorithm was first published by
\insertCite{FineClar82;textual}{fastbeta}.
}

\subsection{S method}{

The susceptible population size is estimated recursively starting from
the supplied initial value \mjseqn{S_0} (\code{par_list$S0}):

\mjsdeqn{S_{i+1} = S_i + B_{i+1} - Z_{i+1} - \mu_i S_i \Delta t\,.}

The infectious population size is approximated by a scaling of incidence:

\mjsdeqn{I_i = \frac{Z_{i-g+1}}{(\gamma + \mu_i) \Delta t}\,,\quad \gamma = 1 / t_\text{gen}\,,}

where \mjseqn{g = \mathrm{nint}(t_\text{gen} / \Delta t)} is the mean
generation interval \mjseqn{t_\text{gen}} in units of the observation
interval \mjseqn{\Delta t} (\code{par_list$tgen}), rounded to the nearest
integer. The transmission rate per susceptible individual
per infectious individual is then estimated as

\mjsdeqn{\beta_i = \frac{Z_{i+1}}{S_i I_i \Delta t}\,.}
}

\subsection{SI method}{

The susceptible and infectious population sizes are estimated recursively
starting from the supplied initial values \mjseqn{S_0} (\code{par_list$S0})
and \mjseqn{I_0} (\code{par_list$I0}):

\mjsdeqn{\begin{align*} S_{i+1} &= \frac{\big\lbrack 1 - \frac{1}{2} \mu_i \Delta t \big\rbrack S_i + B_{i+1} - Z_{i+1}}{1 + \frac{1}{2} \mu_{i+1} \Delta t}\,, \cr I_{i+1} &= \frac{\big\lbrack 1 - \frac{1}{2} (\gamma + \mu_i) \Delta t \big\rbrack I_i + Z_{i+1}}{1 + \frac{1}{2} (\gamma + \mu_{i+1}) \Delta t}\,,\quad \gamma = 1 / t_\text{gen}\,. \end{align*}}

The transmission rate per susceptible individual
per infectious individual is then estimated as

\mjsdeqn{\beta_i = \frac{Z_i + Z_{i+1}}{2 S_i I_i \Delta t}\,.}
}

\subsection{SEI method}{

The susceptible, exposed (infected but not infectious), and infectious
population sizes are estimated recursively starting from the supplied
initial values \mjseqn{S_0} (\code{par_list$S0}), \mjseqn{E_0} (\code{par_list$E0}),
and \mjseqn{I_0} (\code{par_list$I0}):

\mjsdeqn{\begin{align*} S_{i+1} &= \frac{\big\lbrack 1 - \frac{1}{2} \mu_i \Delta t \big\rbrack S_i + B_{i+1} - Z_{i+1}}{1 + \frac{1}{2} \mu_{i+1} \Delta t}\,, \cr E_{i+1} &= \frac{\big\lbrack 1 - \frac{1}{2} (\sigma + \mu_i) \Delta t \big\rbrack E_i + Z_i}{1 + \frac{1}{2} (\sigma + \mu_{i+1}) \Delta t}\,,\quad \sigma = 1 / t_\text{lat}\,, \cr I_{i+1} &= \frac{\big\lbrack 1 - \frac{1}{2} (\gamma + \mu_i) \Delta t \big\rbrack I_i + \frac{1}{2} \sigma (E_i + E_{i+1}) \Delta t}{1 + \frac{1}{2} (\gamma + \mu_{i+1}) \Delta t}\,,\quad \sigma = 1 / t_\text{lat}\,,\quad \gamma = 1 / t_\text{inf}\,. \end{align*}}

The transmission rate per susceptible individual
per infectious individual is then estimated as

\mjsdeqn{\beta_i = \frac{Z_i + Z_{i+1}}{2 S_i I_i \Delta t}\,.}
}

}

\subsection{2. Missing data and zeros in incidence}{

Missing values in \code{data$Z} are not tolerated and must be imputed.
Columns \code{S} and \code{beta} in the output will be filled with \code{NA}
after the index of the first missing value. In the FC and S methods,
zeros in \code{data$Z} can cause divide-by-zero errors. In this case,
column \code{beta} in the output may contain \code{NaN} and \code{Inf} where an
estimate would otherwise be expected. In the SI and SEI methods,
strings of zeros in \code{data$Z} can lead to spurious zeros and large
numeric elements in column \code{beta} of the output. Zeros in \code{data$Z}
can be "imputed" (replaced with positive values in a sensible way)
to avoid these issues. \code{estimate_beta_method()} does not attempt
imputation. However, see \code{\link[=fastbeta]{fastbeta()}}.
}

\subsection{3. Negative susceptibles}{

If true incidence is systematically overestimated by \code{data$Z}
or true births are systematically underestimated by \code{data$B},
then the estimated susceptible population size can grossly
underestimate the true number of susceptibles and can even
become negative. In the latter case, column \code{S} in the output
will contain negative elements. In both cases, more reasonable
results are typically obtained by restarting with scaled up
\code{data$Z} and/or scaled down \code{data$B}. \code{estimate_beta_method()}
does not warn about negative susceptibles. However, see \code{\link[=fastbeta]{fastbeta()}}.
}

\subsection{4. Noise in the estimated transmission rate}{

All four estimation methods propagate noise (due to process
and observation error) from the data to transmission rate
estimates. To distill temporal patterns of interest from the
noise, the raw transmission rate estimates can be smoothed
using, for example, \code{\link[stats:loess]{stats::loess()}}. \code{estimate_beta_method()}
does not undertake smoothing.
}

\subsection{5. Effective observation interval (FC method only)}{

In the absence of errors, every \code{round(par_list$tgen)}th row in
the FC method output will be complete. The remaining rows will
contain \code{NA} in columns \code{S}, \code{I}, \code{beta}, \code{Z_agg}, and \code{B_agg}.
This is not a mistake: the rounded generation interval becomes
the \emph{effective} observation interval when the incidence and births
time series are aggregated (see Algorithm).
}
}

\examples{
# Simulate time series data using an SIR model
# to test the FC, S, and SI methods
pl_sir <- make_par_list(model = "sir")
df_sir <- make_data(pl_sir, with_ds = FALSE, model = "sir")
head(df_sir)

# Simulate time series data using an SEIR model
# to test the SEI method
pl_seir <- make_par_list(model = "seir")
df_seir <- make_data(pl_seir, with_ds = FALSE, model = "seir")
head(df_seir)

# Estimate the seasonally forced transmission rate
# underlying each simulation
f_list <- list(
  fc  = estimate_beta_fc,
  s   = estimate_beta_s,
  si  = estimate_beta_si,
  sei = estimate_beta_sei
)
eb_out <- mapply(function(f, df, pl) f(data = df, par_list = pl),
  f = f_list,
  df = list(df_sir, df_sir, df_sir, df_seir),
  pl = list(pl_sir, pl_sir, pl_sir, pl_seir),
  SIMPLIFY = FALSE
)
lapply(eb_out, head)

op <- par(mfrow = c(3, 1), mar = c(2, 6, 1, 1), oma = c(3, 0, 2, 0))

# FC method fails rapidly as a result of ignoring
# susceptible mortality
plot(S ~ t, df_sir, type = "l", lwd = 4, ylim = c(43, 83) * 1e03,
     xlab = "", ylab = "Susceptible")
lines(S ~ t, na.omit(eb_out$fc), lwd = 2, col = "seagreen")
title(main = "FC method", line = 1, xpd = NA)
plot(I ~ t, df_sir, type = "l", lwd = 4, ylim = c(0, 3.5) * 1e03,
     xlab = "", ylab = "Infectious")
lines(I ~ t, na.omit(eb_out$fc), lwd = 2, col = "mediumvioletred")
plot(beta ~ t, df_sir, type = "l", lwd = 4, ylim = c(0.5, 1.3) * 1e-05,
     xlab = "", ylab = "Transmission rate")
lines(beta ~ t, na.omit(eb_out$fc), lwd = 2, col = "slateblue")
title(xlab = "Time (units dt)", line = 3, xpd = NA)

# S method is visibly biased
plot(S ~ t, df_sir, type = "l", lwd = 4, ylim = c(43, 58) * 1e03,
     xlab = "", ylab = "Susceptible")
lines(S ~ t, eb_out$s, lwd = 2, col = "seagreen")
title(main = "S method", line = 1, xpd = NA)
plot(I ~ t, df_sir, type = "l", lwd = 4, ylim = c(0, 3.5) * 1e03,
     xlab = "", ylab = "Infectious")
lines(I ~ t, eb_out$s, lwd = 2, col = "mediumvioletred")
plot(I(beta / pl_sir$beta_mean) ~ t, df_sir, type = "l", lwd = 4, ylim = c(0.7, 1.3),
     xlab = "", ylab = "Transmission rate\nrelative to mean")
lines(I(beta / pl_sir$beta_mean) ~ t, eb_out$s, lwd = 2, col = "slateblue")
title(xlab = "Time (units dt)", line = 3, xpd = NA)

# SI method does well
plot(S ~ t, df_sir, type = "l", lwd = 4, ylim = c(43, 58) * 1e03,
     xlab = "", ylab = "Susceptible")
lines(S ~ t, eb_out$si, lwd = 2, col = "seagreen")
title(main = "SI method", line = 1, xpd = NA)
plot(I ~ t, df_sir, type = "l", lwd = 4, ylim = c(0, 3.5) * 1e03,
     xlab = "", ylab = "Infectious")
lines(I ~ t, eb_out$si, lwd = 2, col = "mediumvioletred")
plot(I(beta / pl_sir$beta_mean) ~ t, df_sir, type = "l", lwd = 4, ylim = c(0.7, 1.3),
     xlab = "", ylab = "Transmission rate\nrelative to mean")
lines(I(beta / pl_sir$beta_mean) ~ t, eb_out$si, lwd = 2, col = "slateblue")
title(xlab = "Time (units dt)", line = 3, xpd = NA)

# SEI method does well
plot(S ~ t, df_seir, type = "l", lwd = 4, ylim = c(43, 58) * 1e03,
     xlab = "", ylab = "Susceptible")
lines(S ~ t, eb_out$sei, lwd = 2, col = "seagreen")
title(main = "SEI method", line = 1, xpd = NA)
plot(I ~ t, df_seir, type = "l", lwd = 4, ylim = c(0, 2) * 1e03,
     xlab = "", ylab = "Infectious")
lines(I ~ t, eb_out$sei, lwd = 2, col = "mediumvioletred")
plot(I(beta / pl_seir$beta_mean) ~ t, df_seir, type = "l", lwd = 4, ylim = c(0.7, 1.3),
     xlab = "", ylab = "Transmission rate\nrelative to mean")
lines(I(beta / pl_seir$beta_mean) ~ t, eb_out$sei, lwd = 2, col = "slateblue")
title(xlab = "Time (units dt)", line = 3, xpd = NA)

par(op)

# NOTE: Restarting with `with_ds = TRUE` passed to
# `make_data()` suggests that the SI method is the
# most robust to demographic stochasticity and more
# generally to noise in the data-generating process.

}
\references{
\insertRef{Jaga+20}{fastbeta}
\insertRef{FineClar82}{fastbeta}
}
\seealso{
\code{\link[=fastbeta]{fastbeta()}}
}
