% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_beta_FC.R
\name{estimate_beta_FC}
\alias{estimate_beta_FC}
\title{Estimate time-varying transmission rates (FC method)}
\usage{
estimate_beta_FC(df = data.frame(), par_list = list())
}
\arguments{
\item{df}{A data frame with numeric columns:

\describe{
\item{\code{t}}{Time. \code{t[i]} is equal to
\ifelse{latex}{\out{$t_{i-1} = t_0 + (i-1) \Delta t$}}{\ifelse{html}{\out{<i>t</i><sub><i>i</i>&minus;1</sub> = <i>t</i><sub>0</sub>+(<i>i</i>&minus;1)<i>&Delta;t</i>}}{t_i-1 = t_0 + (i-1)*Dt}}
in units
\ifelse{latex}{\out{$\Delta t$}}{\ifelse{html}{\out{<i>&Delta;t</i>}}{Dt}},
so that \code{t[i] - t[i-1]} is equal to 1.
}
\item{\code{C}}{Reported incidence. \code{C[i]} is the number of cases
reported between times \code{t[i-1]} and \code{t[i]}.
}
\item{\code{B}}{Births. \code{B[i]} is the number of births between times
\code{t[i-1]} and \code{t[i]}.
}
}

\code{B} is optional if \code{hatN0} and \code{nu} are defined in \code{par_list}
(see Details).}

\item{par_list}{A list containing:

\describe{
\item{\code{prep}}{[ \ifelse{latex}{\out{$p_\text{rep}$}}{\ifelse{html}{\out{<i>p</i><sub>rep</sub>}}{p_rep}} ]
Case reporting probability.
}
\item{\code{trep}}{[ \ifelse{latex}{\out{$t_\text{rep}$}}{\ifelse{html}{\out{<i>t</i><sub>rep</sub>}}{t_rep}} ]
Case reporting delay in units
\ifelse{latex}{\out{$\Delta t$}}{\ifelse{html}{\out{<i>&Delta;t</i>}}{Dt}}.
}
\item{\code{S0}}{[ \ifelse{latex}{\out{$S_0$}}{\ifelse{html}{\out{<i>S</i><sub>0</sub>}}{S_0}} ]
Number of susceptibles at time
\ifelse{latex}{\out{$t = t_0$}}{\ifelse{html}{\out{<i>t</i> = <i>t</i><sub>0</sub>}}{t = t_0}}.
}
\item{\code{hatN0}}{[ \ifelse{latex}{\out{$\widehat{N}_0$}}{\ifelse{html}{\out{<i>&Ntilde;</i><sub>0</sub>}}{hatN_0}} ]
Population size at time
\ifelse{latex}{\out{$t = 0$}}{\ifelse{html}{\out{<i>t</i> = 0}}{t = 0}}
years.
}
\item{\code{nu}}{[ \ifelse{latex}{\out{$\nu_\text{c}$}}{\ifelse{html}{\out{<i>&nu;<sub>c</sub></i>}}{nu_c}} ]
Birth rate expressed per unit
\ifelse{latex}{\out{$\Delta t$}}{\ifelse{html}{\out{<i>&Delta;t</i>}}{Dt}}
and relative to
\ifelse{latex}{\out{$\hat{N}_0$}}{\ifelse{html}{\out{<i>&Ntilde;</i><sub>0</sub>}}{hatN_0}}
(if modeled as constant).
}
\item{\code{tgen}}{[ \ifelse{latex}{\out{$t_\text{gen}$}}{\ifelse{html}{\out{<i>t</i><sub>gen</sub>}}{t_gen}} ]
Mean generation interval of the disease of interest in units
\ifelse{latex}{\out{$\Delta t$}}{\ifelse{html}{\out{<i>&Delta;t</i>}}{Dt}}.
}
}

\code{hatN0} and \code{nu} are optional if \code{B} is defined in \code{df}
(see Details).}
}
\value{
A data frame with numeric columns:

\describe{
\item{\code{t}}{Time. Identical to \code{df$t}.}
\item{\code{C}}{Reported incidence, imputed. Identical to \code{df$C},
except with missing values and zeros imputed (see Details).
}
\item{\code{Z}}{Incidence. \code{Z[i]} is the estimated
number of infections between times \code{t[i-1]} and \code{t[i]}.
}
\item{\code{Z_agg}}{Aggregated incidence. \code{Z_agg[i]} is the estimated
number of infections between times \code{t[i-round(tgen)+1]} and
\code{t[i]}, for \code{i} in \code{seq(1 + round(tgen), nrow(df), by = round(tgen))}.
\code{Z_agg[i]} is \code{NA} otherwise.
}
\item{\code{B}}{Births, imputed. Identical to \code{df$B} (if supplied),
except with missing values imputed (see Details).
}
\item{\code{B_agg}}{Aggregated births. \code{B_agg[i]} is the
number of births between times \code{t[i-round(tgen)+1]} and \code{t[i]},
for \code{i} in \code{seq(1 + round(tgen), nrow(df), by = round(tgen))}.
\code{B_agg[i]} is \code{NA} otherwise.
}
\item{\code{S}}{Number of susceptibles. \code{S[i]} is the estimated
number of susceptibles at time \code{t[i]}.
}
\item{\code{I}}{Number of infecteds. \code{I[i]} is the estimated
number of infecteds at time \code{t[i]}.
}
\item{\code{beta}}{Transmission rate. \code{beta[i]} is the estimated
transmission rate at time \code{t[i]} expressed per unit
\ifelse{latex}{\out{$\Delta t$}}{\ifelse{html}{\out{<i>&Delta;t</i>}}{Dt}}
per susceptible per infected.
}
}

It possesses \code{par_list} as an attribute.
}
\description{
\code{estimate_beta_FC()} applies the FC method (see References)
to estimate the time-varying transmission rate
\ifelse{latex}{\out{$\beta(t)$}}{\ifelse{html}{\out{<i>&beta;</i>(<i>t</i>)}}{beta(t)}}
from time series of reported incidence and births, observed at
equally spaced time points
\ifelse{latex}{\out{$t_k = t_0 + k \Delta t$}}{\ifelse{html}{\out{<i>t<sub>k</sub></i> = <i>t</i><sub>0</sub>+<i>k&Delta;t</i>}}{t_k = t_0 + k*Dt}}
(for \ifelse{latex}{\out{$k = 0,\ldots,n$}}{\ifelse{html}{\out{<i>k</i> = 0,...,<i>n</i>}}{k = 0,...,n}}),
where
\ifelse{latex}{\out{$\Delta t$}}{\ifelse{html}{\out{<i>&Delta;t</i>}}{Dt}}
denotes the observation interval.
}
\details{
In the output, every \code{with(par_list, round(tgen))} rows are
complete. The remaining rows contain \code{NA} in columns \code{S}, \code{I},
\code{beta}, \code{Z_agg}, and \code{B_agg}. This is a limitation of the FC
method, which aggregates incidence and births over the mean
generation interval, making \code{with(par_list, round(tgen))} the
\emph{effective} observation interval. If this number is zero, then
the method fails.

If \code{df$B} is undefined in the function call, then \code{df$B[i]}
gets the value \code{with(par_list, nu * hatN0 * 1)} for all \code{i}.

Missing values in \code{df[, c("C", "B")]} are not tolerated by
the FC method. They are imputed via linear interpolation
between observed values. If there are no observations
before the first missing value, then complete imputation
is impossible. In this case, the FC method may fail:
columns \code{S} and \code{beta} in the output may be filled with \code{NA}.

Zeros in \code{df$C} cause divide-by-zero errors. To prevent these
errors, zeros are imputed like missing values. If there are
no nonzero observations before the first zero, then complete
imputation is impossible. In this case, the FC method may fail,
but only locally: column \code{beta} in the output may contain some
\code{NaN} and \code{Inf}, but give estimates everywhere else.
}
\references{
deJonge MS, Jagan M, Krylova O, Earn DJD. Fast estimation of
time-varying transmission rates for infectious diseases.
}
