% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastbeta-package.R
\docType{package}
\name{fastbeta-package}
\alias{fastbeta-package}
\alias{_PACKAGE}
\title{fastbeta: Fast Estimation of Time-Varying Transmission Rates}
\description{
Fast methods for estimating time-varying infectious disease
    transmission rates from incidence and mortality time series.
}
\details{
\loadmathjax
\pkg{fastbeta} implements the FC, S, and SI methods for estimating
time-varying infectious disease transmission rates \mjseqn{\beta(t)}
from disease incidence and mortality data:
\itemize{
\item \code{\link[=estimate_beta_fc]{estimate_beta_fc()}}
\item \code{\link[=estimate_beta_s]{estimate_beta_s()}}
\item \code{\link[=estimate_beta_si]{estimate_beta_si()}}
}

The SI method is substantially more robust than the FC and S methods
and should be preferred in practice.

\pkg{fastbeta} additionally implements peak-to-peak iteration
(PTPI), a method for estimating the initial number of susceptibles
\mjseqn{S_0} from time series data:
\itemize{
\item \code{\link[=ptpi]{ptpi()}}
\item \code{\link[=get_peak_times]{get_peak_times()}}
}

PTPI can be used in conjunction with the SI method, which requires
users to specify an estimate of \mjseqn{S_0}.

\pkg{fastbeta} includes functions useful for simulating incidence
time series with an underlying, seasonally forced transmission rate.
These may be helpful in testing:
\itemize{
\item \code{\link[=make_par_list]{make_par_list()}}
\item \code{\link[=make_data]{make_data()}}
\item \code{\link[=compute_rrmse]{compute_rrmse()}}
}

All methods are based on the SIR model with time-varying rates of
birth, death, and transmission:

\mjsdeqn{\begin{align} \frac{\text{d}S}{\text{d}t} &= \nu(t) \widehat{N}_0 - \beta(t) S I - \mu(t) S \cr \frac{\text{d}I}{\text{d}t} &= \beta(t) S I - \gamma I - \mu(t) I \cr \frac{\text{d}R}{\text{d}t} &= \gamma I - \mu(t) R \end{align}}

See References for details.
}
\references{
\insertRef{Jaga+20}{fastbeta}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/davidearn/fastbeta}
}

}
\author{
\strong{Maintainer}: Mikael Jagan \email{mjagan@uwaterloo.ca}

}
\keyword{internal}
