\name{fastbeta-package}
\alias{fastbeta-package}
\title{\R{} Package \pkg{fastbeta}}
\description{
An \R{} package for approximating time-varying infectious disease
transmission rates from disease incidence time series and other data.
}
\details{
The \dQuote{main} function is \code{\link{fastbeta}}.

To render a list of available help topics, use
\code{\link{help}(package = "fastbeta")}.

To report a bug or request a change, use
\code{\link{bug.report}(package = "fastbeta")}.
}
\author{
Mikael Jagan
\email{jaganmn@mcmaster.ca}
}
\docType{package}
