% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastbeta.R
\name{fastbeta}
\alias{fastbeta}
\title{\loadmathjax
Estimate time-varying transmission rates}
\usage{
fastbeta(data, par_list, method)
}
\arguments{
\item{data}{A data frame with numeric columns:

\describe{
\item{\code{t}}{\mjseqn{\lbrace\,t_i\,\rbrace}
Time. Lists the observation times \mjseqn{t_i = t_0 + i \Delta t}
(for \mjseqn{i = 0, \ldots, n-1}) in any convenient units. Here,
\mjseqn{t_0} is the initial observation time and \mjseqn{\Delta t}
is the observation interval.
}
\item{\code{Z}}{\mjseqn{\lbrace\,Z_i\,\rbrace}
Incidence. \code{Z[i]} is the number of infections
between times \code{t[i-1]} and \code{t[i]}.
}
\item{\code{B}}{\mjseqn{\lbrace\,B_i\,\rbrace}
Births. \code{B[i]} is the number of births
between times \code{t[i-1]} and \code{t[i]}.
}
\item{\code{mu}}{\mjseqn{\lbrace\,\mu_i \Delta t\,\rbrace}
Per capita natural mortality rate. \code{mu[i]} is the rate
at time \code{t[i]} expressed per unit \mjseqn{\Delta t}.
S, SI, and SEI methods only.
}
}}

\item{par_list}{A list with numeric scalar elements:

\describe{
\item{\code{tgen}}{\mjseqn{\lbrace\,t_\text{gen} / \Delta t\,\rbrace}
Mean generation interval of the disease of interest
in units \mjseqn{\Delta t}. FC, S, and SI methods only.
}
\item{\code{tlat}}{\mjseqn{\lbrace\,t_\text{lat} / \Delta t\,\rbrace}
Mean latent period of the disease of interest
in units \mjseqn{\Delta t}. SEI method only.
}
\item{\code{tinf}}{\mjseqn{\lbrace\,t_\text{inf} / \Delta t\,\rbrace}
Mean infectious period of the disease of interest
in units \mjseqn{\Delta t}. SEI method only.
}
\item{\code{S0}}{\mjseqn{\lbrace\,S_0\,\rbrace}
Number of susceptible individuals at time \mjseqn{t = t_0}.
}
\item{\code{E0}}{\mjseqn{\lbrace\,E_0\,\rbrace}
Number of exposed (infected but not infectious) individuals
at time \mjseqn{t = t_0}. SEI method only.
}
\item{\code{I0}}{\mjseqn{\lbrace\,I_0\,\rbrace}
Number of infectious individuals at time \mjseqn{t = t_0}.
SI and SEI methods only.
}
}}

\item{method}{A character scalar, one of \code{"fc"}, \code{"s"}, \code{"si"}, and \code{"sei"},
indicating an estimation algorithm, one of \code{\link[=estimate_beta_fc]{estimate_beta_fc()}},
\code{\link[=estimate_beta_s]{estimate_beta_s()}}, \code{\link[=estimate_beta_si]{estimate_beta_si()}}, and \code{\link[=estimate_beta_sei]{estimate_beta_sei()}}.}
}
\value{
A fastbeta object. A list with elements:

\describe{
\item{\code{out}}{A data frame containing the estimation output.
The result of applying \code{\link[=estimate-beta]{estimate_beta_method()}}
to \code{data} and \code{par_list} \emph{after} replacing missing values in
\code{data[c("Z", "B", "mu")]} and zeros in \code{data$Z} with \code{\link[=impute_na]{impute_na()}}
(see Details).
}
\item{\code{method}}{A character scalar. The value of \code{method}
in the function call.
}
\item{\code{warn}}{A logical scalar indicating whether a warning
was issued about negative elements in \code{out$S} (see Details).
Equal to \code{any(out$S < 0, na.rm = TRUE)}.
}
\item{\code{call}}{The function call. The fastbeta object is reproducible
with \code{eval(call)}.
}
\item{\code{arg_list}}{A list of the arguments in the function call.
The fastbeta object is reproducible with \code{do.call(fastbeta, arg_list)}.
}
}
}
\description{
A wrapper for the \code{\link[=estimate-beta]{estimate_beta_method()}}
functions. These implement the FC, S, SI, and SEI methods for
estimating time-varying transmission rates \mjseqn{\beta(t)}
from time series of incidence, births, and natural mortality
observed at equally spaced time points \mjseqn{t_i = t_0 + i \Delta t}.
Note that the algorithms require incidence, \emph{not} reported
incidence. See \code{\link[=deconvol]{deconvol()}} for a method of reconstructing
incidence from reported incidence.
}
\section{Details}{
\subsection{1. Models and algorithms}{

See \link[=estimate-beta]{here}.
}

\subsection{2. Missing data and zeros in incidence}{

Missing values in \code{data} are not tolerated by any of the estimation
methods. Columns \code{S} and \code{beta} in the output will be filled with \code{NA}
after the index of the first encountered missing value.

In the FC and S methods, zeros in \code{data$Z} can cause divide-by-zero
errors. In this case, column \code{beta} in the output may contain \code{NaN}
and \code{Inf} where an estimate would otherwise be expected. In the SI
and SEI methods, strings of zeros in \code{data$Z} can lead to spurious
zeros and large numeric elements in column \code{beta} of the output.

\code{fastbeta()} applies \code{\link[=impute_na]{impute_na()}} to the columns of \code{data} in an
attempt to impute \code{NA} in \code{data} and replace zeros in \code{data$Z} with
positive values. \code{\link[=impute_na]{impute_na()}} fails for missing values and zeros
at the edges of a vector. If a different imputation scheme is
desired (\code{\link[=impute_na]{impute_na()}} uses linear interpolation), then it must be
applied prior to calling \code{fastbeta()}.
}

\subsection{3. Negative susceptibles}{

If true incidence is systematically overestimated by \code{data$Z}
or true births are systematically underestimated by \code{data$B},
then the estimated susceptible population size can grossly
underestimate the true number of susceptibles and can even
become negative. In the latter case, column \code{S} in the output
will have negative elements.

\code{fastbeta()} warns if the estimated susceptible population size
is at any time negative and suggests restarting with scaled down
\code{data$Z} and/or scaled up \code{data$B}. A rule-of-thumb is to iterate
with different scalings until the susceptible time series no
longer displays pronounced transient dynamics, while considering
whether the correction applied is reasonable.
}

\subsection{4. Noise in the estimated transmission rate}{

All four estimation methods allow noise in the data (due to
process and observation error) to be propagated to transmission
rate estimates. To distill temporal  patterns of interest from
the noise, the raw transmission rate estimates can be smoothed
using, for example, \code{\link[stats:loess]{stats::loess()}}. \code{fastbeta()} does not
undertake smoothing, because determining what constitutes a
"good" degree of smoothing is not a trivial task and typically
involves a graphical inspection of results for different values
of a smoothing parameter. \code{\link[=try_loess]{try_loess()}} can be used to greatly
facilitate the selection of loess fits to noisy transmission
rate estimates.
}

\subsection{5. Confidence intervals on the estimated transmission rate}{

\code{fastbeta()} does not construct confidence intervals on any
of its estimates. The returned fastbeta object can be passed
to \code{\link[=bootbeta]{bootbeta()}} to obtain bootstrap 95\% confidence intervals.
}

\subsection{6. Effective observation interval (FC method only)}{

In the absence of errors, every \code{round(par_list$tgen)}th row in
the FC method output will be complete. The remaining rows will
contain \code{NA} in columns \code{S}, \code{I}, \code{beta}, \code{Z_agg}, and \code{B_agg}.
This is not a mistake: the rounded generation interval becomes
the \emph{effective} observation interval when incidence and births
are aggregated over the rounded generation interval (see the
FC method algorithm \link[=estimate-beta]{here}).
}
}

\examples{
## Simulate time series data using an SIR model
## with seasonally forced transmission rate
pl <- make_par_list(model = "sir")
df <- make_data(pl, with_ds = TRUE, model = "sir")

## Estimate the seasonally forced transmission rate
## using the SI method
fastbeta_out <- fastbeta(df, pl, method = "si")
plot(fastbeta_out)

## Simulate time series data using an SEIR model
## with seasonally forced transmission rate
pl <- make_par_list(model = "seir")
df <- make_data(pl, with_ds = TRUE, model = "seir")

## Estimate the seasonally forced transmission rate
## using the SEI method
fastbeta_out <- fastbeta(df, pl, method = "sei")
plot(fastbeta_out)

}
\references{
\insertRef{Jaga+20}{fastbeta}
}
\seealso{
\link[=fastbeta-methods]{methods for class "fastbeta"},
\link[=estimate-beta]{algorithms called under the hood},
\code{\link[=try_loess]{try_loess()}} for smoothing noisy transmission rate estimates,
\code{\link[=bootbeta]{bootbeta()}} for constructing bootstrap confidence intervals
on transmission rate estimates
}
