% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastbeta.R
\name{fastbeta}
\alias{fastbeta}
\title{Estimate time-varying transmission rates}
\usage{
fastbeta(df, par_list)
}
\arguments{
\item{df}{A data frame with time series data.}

\item{par_list}{A list of parameter values.}
}
\value{
A fastbeta object.
}
\description{
Generates a fastbeta object.
}
\details{
Details to come.
}
\references{
\insertRef{Jaga+20}{fastbeta}
}
\seealso{
\code{\link[=estimate_beta_si]{estimate_beta_si()}}, \code{\link[=deconvolve]{deconvolve()}}, \code{\link[=plot.fastbeta]{plot.fastbeta()}}
}
