% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fastbeta.R
\name{fastbeta}
\alias{fastbeta}
\title{Estimate time-varying transmission rates}
\usage{
fastbeta(df, par_list)
}
\arguments{
\item{df}{A data frame with time series data.}

\item{par_list}{A list of parameter values.}
}
\value{
A fastbeta object.
}
\description{
Generates a fastbeta object.
}
\details{
Details to come.
}
\examples{
# Simulate time series data using an SIR model
# with seasonally forced transmission rate
par_list <- make_par_list(dt_weeks = 1)
df <- make_data(par_list = par_list, with_ds = TRUE)

# Estimate susceptibles, infecteds, and
# the seasonally forced transmission rate
fastbeta_out <- fastbeta(df, par_list)
plot(fastbeta_out)

}
\references{
\insertRef{Jaga+20}{fastbeta}
}
\seealso{
\code{\link[=estimate_beta_si]{estimate_beta_si()}}, \code{\link[=deconvolve]{deconvolve()}}, \code{\link[=plot.fastbeta]{plot.fastbeta()}}
}
