\name{fastbeta}
\alias{fastbeta}
\title{Estimate a time-varying transmission rate}
\description{}
\usage{
fastbeta(cases, births, mu, gamma, S0, I0)
}
\arguments{
  \item{cases}{}
  \item{births}{}
  \item{mu}{}
  \item{gamma}{}
  \item{S0}{}
  \item{I0}{}
}
\details{}
\value{}
\seealso{}
\references{}
\examples{}
