% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_data.R
\name{make_data}
\alias{make_data}
\title{\loadmathjax
Simulate epidemic time series data}
\usage{
make_data(
  par_list = make_par_list(),
  n = 20 * 365/7,
  with_dem_stoch = TRUE,
  ode_control = list(method = "lsoda", rtol = 1e-06, atol = 1e-06)
)
}
\arguments{
\item{par_list}{A list of parameter values containing:
\describe{
\item{\code{dt_weeks}}{\mjseqn{\lbrack \Delta t \rbrack}
Numeric scalar. Observation interval in weeks.
}
\item{\code{t0}}{\mjseqn{\lbrack t_0 \rbrack}
Numeric scalar. Time of the first observation
in units \mjseqn{\Delta t}.
}
\item{\code{prep}}{\mjseqn{\lbrack p_\text{rep} \rbrack}
Numeric scalar. Case reporting probability.
}
\item{\code{trep}}{\mjseqn{\lbrack t_\text{rep} \rbrack}
Numeric scalar. Case reporting delay in units
\mjseqn{\Delta t}.
}
\item{\code{hatN0}}{\mjseqn{\lbrack \out{\widehat{N}_0} \rbrack}
Numeric scalar. Population size at time \mjseqn{t = 0} years.
}
\item{\code{N0}}{\mjseqn{\lbrack N_0 \rbrack}
Numeric scalar. Population size at time \mjseqn{t = t_0}.
}
\item{\code{S0}}{\mjseqn{\lbrack S_0 \rbrack}
Numeric scalar. Number of susceptibles at time \mjseqn{t = t_0}.
}
\item{\code{I0}}{\mjseqn{\lbrack I_0 \rbrack}
Numeric scalar. Number of infecteds at time \mjseqn{t = t_0}.
}
\item{\code{nu}}{\mjseqn{\lbrack \nu_\text{c} \rbrack}
Numeric scalar. Birth rate expressed
per unit \mjseqn{\Delta t} and relative to
\mjseqn{\out{\widehat{N}_0}}.
}
\item{\code{mu}}{\mjseqn{\lbrack \mu_\text{c} \rbrack}
Numeric scalar. Natural mortality rate expressed
per unit \mjseqn{\Delta t} and per capita.
}
\item{\code{tgen}}{\mjseqn{\lbrack t_\text{gen} \rbrack}
Numeric scalar. Mean generation interval of the disease
of interest in units \mjseqn{\Delta t}.
}
\item{\code{beta_mean}}{\mjseqn{\lbrack \langle\beta\rangle \rbrack}
Numeric scalar. Mean (long-term average) of the seasonally
forced transmission rate \mjseqn{\beta(t)} expressed per unit
\mjseqn{\Delta t} per susceptible per infected.
}
\item{\code{alpha}}{\mjseqn{\lbrack \alpha \rbrack}
Numeric scalar. Amplitude of the seasonally forced transmission
rate \mjseqn{\beta(t)} relative to the mean.
}
\item{\code{epsilon}}{\mjseqn{\lbrack \epsilon \rbrack}
Numeric scalar. Standard deviation of the random phase shift
introduced to the seasonally forced transmission rate
\mjseqn{\beta(t)}.
}
}}

\item{n}{Integer scalar. Time between the first and last observations
in units \mjseqn{\Delta t}, so that simulated time series have \code{n + 1}
observations. (If numeric but not integer, then \code{n} is replaced by
\code{floor(n)}.)}

\item{with_dem_stoch}{Logical scalar. If \code{TRUE}, then the simulation
is generated using \code{\link[adaptivetau:ssa.adaptivetau]{adaptivetau::ssa.adaptivetau()}}. Otherwise, it
is generated using \code{\link[deSolve:ode]{deSolve::ode()}} (see Details).}

\item{ode_control}{A list of optional arguments of \code{\link[deSolve:ode]{deSolve::ode()}},
specifying options for numerical integration, such as \code{method},
\code{rtol}, and \code{atol}. Not used if \code{with_dem_stoch = TRUE} (see Details).}
}
\value{
A data frame with \code{n + 1} rows corresponding to equally spaced times
\mjseqn{t_k = t_0 + k \Delta t} (for \mjseqn{k = 0, \ldots, n}), and
numeric columns:

\describe{
\item{\code{t}}{Time in units \mjseqn{\Delta t}. Equal to \code{par_list$t0 + 0:n}.}
\item{\code{t_years}}{Time in years. Equal to
\code{(par_list$t0 + 0:n) * par_list$dt_weeks * (7 / 365)}.
}
\item{\code{beta}}{Seasonally forced transmission rate expressed
per unit \mjseqn{\Delta t} per susceptible per infected,
without environmental noise.
}
\item{\code{beta_phi}}{Seasonally forced transmission rate expressed
per unit \mjseqn{\Delta t}{dt} per susceptible per infected,
with environmental noise.
}
\item{\code{N}}{Population size.}
\item{\code{S}}{Number of susceptibles.}
\item{\code{I}}{Number of infecteds.}
\item{\code{Z}}{Incidence. \code{Z[i]} is the number of infections
between times \code{t[i-1]} and \code{t[i]}.
}
\item{\code{C}}{Reported incidence. \code{C[i]} is the number of infections
reported between times \code{t[i-1]} and \code{t[i]}, equal to the number
of successes in \code{Z[i-round(par_list$trep)]} independent Bernoulli
trials, each with success probability \code{par_list$prep}.
}
\item{\code{B}}{Births. \code{B[i]} is the number of births between times
\code{t[i-1]} and \code{t[i]}.
}
\item{\code{mu}}{Per capita natural mortality rate. \code{mu[i]} is rate
at time \code{t[i]}. Equal to \code{rep(par_list$mu, n + 1)} and included
for convenience, mainly because \code{\link[=estimate_beta_si]{estimate_beta_si()}} requires
a data frame with columns \code{Z}, \code{B}, and \code{mu}.
}
}

The data frame has attributes \code{call} and \code{arg_list}, making it
reproducible with \code{eval(call)} or \code{do.call(make_data, arg_list)},
preceded by a call to \code{set.seed()}.
}
\description{
\code{make_data()} simulates epidemic time series data using a system of
SIR equations and a supplied list of parameter values. Observations
are recorded at equally spaced time points \mjseqn{t_k = t_0 + k \Delta t}
(for \mjseqn{k = 0, \ldots, n}). Among other things, the simulation model
assumes:
\itemize{
\item A seasonally forced transmission rate.
\item Constant birth and per capita natural mortality rates.
\item Cases reported after a fixed delay since infection,
with a fixed probability.
}
}
\section{Details}{
\subsection{Simulation model}{

\code{make_data()} simulates epidemic time series data using the system of
SIR equations below, which includes a fourth equation for cumulative
births and a fifth equation for cumulative incidence:

\mjseqn{\out{\begin{align} \frac{\text{d}S}{\text{d}t} &= \nu \widehat{N}_0 - \beta_\phi(t) S I - \mu S\,, \cr \frac{\text{d}I}{\text{d}t} &= \beta_\phi(t) S I - \gamma I - \mu I\,, \cr \frac{\text{d}R}{\text{d}t} &= \gamma I - \mu R\,, \cr \frac{\text{d}B_\text{cum}}{\text{d}t} &= \nu \widehat{N}_0\,, \cr \frac{\text{d}Z_\text{cum}}{\text{d}t} &= \beta_\phi(t) S I\,. \end{align}}}

Here, \mjseqn{\gamma = 1 / t_\text{gen}},

\mjsdeqn{\beta_\phi(t) = \langle\beta\rangle \left\lbrack 1 + \alpha \cos\left(\frac{2 \pi t}{\text{1 year}} + \phi(t;\epsilon)\right) \right\rbrack\,,}

and \mjseqn{\phi(t;\epsilon)} is the linear interpolant of noise
\mjseqn{\lbrace(t_k;\Phi_k)\rbrace_{k=0}^n} with

\mjsdeqn{\Phi_k \sim \mathrm{Normal}(0,\epsilon^2)\,,}

modeling \strong{environmental stochasticity}.

\code{make_data()} generates observations of the above system at equally spaced
times \mjseqn{t_k = t_0 + k \Delta t} (for \mjseqn{k = 0, \ldots, n}) by
either
(i) numerically integrating the ODE using \code{\link[deSolve:ode]{deSolve::ode()}}
(\code{with_dem_stoch = FALSE}), or
(ii) realizing a corresponding continuous-time stochastic process using
\code{\link[adaptivetau:ssa.adaptivetau]{adaptivetau::ssa.adaptivetau()}} (\code{with_dem_stoch = TRUE}).
Both methods use initial state

\mjsdeqn{\begin{bmatrix} S(t_0) \cr I(t_0) \cr R(t_0) \cr B_\text{cum}(t_0) \cr Z_\text{cum}(t_0) \end{bmatrix} = \begin{bmatrix} S_0 \cr I_0 \cr N_0 - S_0 - I_0 \cr 0 \cr 0 \end{bmatrix}\,.}

The latter method defines event probabilities as proportional to
terms in the ODE and models \strong{demographic stochasticity}. Disease
fadeout in simulations with demographic stochasticity is prevented
by setting the probability of infected decrease to zero whenever
the number of infecteds is one.

\code{make_data()} calculates births \mjseqn{B} and incidence \mjseqn{Z}
from cumulative births \mjseqn{B_\text{cum}} and cumulative incidence
\mjseqn{Z_\text{cum}} via first differences,

\mjsdeqn{\begin{align} B(t_k) &= B_\text{cum}(t_k) - B_\text{cum}(t_{k-1}), \cr Z(t_k) &= Z_\text{cum}(t_k) - Z_\text{cum}(t_{k-1}), \end{align}}

then simulates reported incidence \mjseqn{C} from
incidence \mjseqn{Z} via lagged binomial sampling,

\mjsdeqn{C(t_{k+r}) \sim \mathrm{Binomial}\big(Z(t_k),p_\text{rep}\big)\,,}

modeling \strong{observation error} (random under-reporting of cases with
a delay between infection and reporting). Here,
\mjseqn{r = \mathrm{nint}(t_\text{rep} / \Delta t)}.
}
}

\examples{
# Deterministic simulation
par_list <- make_par_list(dt_weeks = 1)
df <- make_data(
  par_list = par_list,
  n = 20 * 365 / 7, # number of weeks in 20 years
  with_dem_stoch = FALSE
)
head(df)

# Stochastic simulation
par_list <- make_par_list(
  dt_weeks = 1,
  epsilon = 0.5, # environmental stochasticity
  prep = 0.5 # random under-reporting of cases
)
set.seed(2146)
df <- make_data(
  par_list = par_list,
  n = 20 * 365 / 7,
  with_dem_stoch = TRUE # demographic stochasticity
)
head(df)

}
