% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_data.R
\name{make_par_list}
\alias{make_par_list}
\title{\loadmathjax
Create a list of parameter values for simulations}
\usage{
make_par_list(
  dt_days = 7,
  N0 = 1e+06,
  tlat = 5/dt_days,
  tinf = 7/dt_days,
  muconst = 0.04 * dt_days/365,
  Rnaught = 20,
  alpha = 0.08,
  epsilon = 0,
  prep = 1,
  model = "seir",
  n = 1000 * 365/dt_days
)
}
\arguments{
\item{dt_days}{\mjseqn{\lbrace\,\Delta t\,\rbrace}
A numeric scalar. Observation interval in days.}

\item{N0}{\mjseqn{\lbrace\,N_0\,\rbrace}
A numeric scalar. Population size at time \mjseqn{t = 0} years.}

\item{tlat}{\mjseqn{\lbrace\,t_\text{lat}/\Delta t\,\rbrace}
A numeric scalar. Mean latent period
of the disease of interest in units \mjseqn{\Delta t}.}

\item{tinf}{\mjseqn{\lbrace\,t_\text{inf}/\Delta t\,\rbrace}
A numeric scalar. Mean infectious period
of the disease of interest in units \mjseqn{\Delta t}.}

\item{muconst}{\mjseqn{\lbrace\,\mu_\text{c} \Delta t\,\rbrace}
A numeric scalar. Birth rate (relative to \mjseqn{N_0})
expressed per unit \mjseqn{\Delta t} \emph{and} per capita
natural mortality rate expressed per unit \mjseqn{\Delta t}.}

\item{Rnaught}{\mjseqn{\lbrace\,\mathcal{R}_0\,\rbrace}
A numeric scalar. Basic reproduction number
of the disease of interest.}

\item{alpha}{\mjseqn{\lbrace\,\alpha\,\rbrace}
A numeric scalar. Amplitude of the seasonally forced
transmission rate relative to its mean.}

\item{epsilon}{\mjseqn{\lbrace\,\epsilon\,\rbrace}
A numeric scalar. Standard deviation of the standard
normally distributed phase shift in the seasonally forced
transmission rate.}

\item{prep}{\mjseqn{\lbrace\,p_\text{rep}\,\rbrace}
A numeric scalar. Probability that an infection is reported.}

\item{model}{A character scalar, either \code{"sir"} or \code{"seir"},
indicating a system of equations to be numerically integrated
(see Details 2).}

\item{n}{\mjseqn{\lbrace\,n\,\rbrace}
An integer scalar (positive). A system of SIR (\code{model = "sir"})
or SEIR (\code{model = "seir"}) equations will be numerically integrated
between times \mjseqn{t = -(n-1) \Delta t} and \mjseqn{t = 0} years
to obtain values for \mjseqn{S_0}, \mjseqn{E_0} (\code{model = "seir"}),
and \mjseqn{I_0} (see Details 2 and 3).}
}
\value{
A list of the arguments in the function call,
excluding \code{model} and \code{n} but including these additional elements:

\describe{
\item{\code{beta_mean}}{\mjseqn{\lbrace\,\langle\beta\rangle \Delta t\,\rbrace}
Mean (long-term average) of the seasonally forced
transmission rate expressed per unit \mjseqn{\Delta t}
per susceptible individual per infectious individual.
}
\item{\code{S0}}{\mjseqn{\lbrace\,S_0\,\rbrace}
Number of susceptible individuals at time \mjseqn{t = 0} years.
}
\item{\code{E0}}{\mjseqn{\lbrace\,E_0\,\rbrace}
Number of exposed individuals at time \mjseqn{t = 0} years.
Included only if \code{model = "seir"}.
}
\item{\code{I0}}{\mjseqn{\lbrace\,I_0\,\rbrace}
Number of infected (\code{model = "sir"}) or infectious (\code{model = "seir"})
individuals at time \mjseqn{t = 0} years.
}
}

This list can be assigned to the argument \code{par_list} of \verb{[make_data()]}
if defaults are accepted for its arguments \code{mu}, \code{nu}, and \code{beta}.
In this case, care must be taken to ensure that the argument \code{model}
of \code{\link[=make_data]{make_data()}} matches that of \code{make_par_list()}.
}
\description{
Creates a list of parameter values that can be assigned
to the argument \code{par_list} of \code{\link[=make_data]{make_data()}} if defaults
are accepted for its arguments \code{mu}, \code{nu}, and \code{beta}.
}
\section{Details}{
\subsection{1. Defining \mjseqn{\langle\beta\rangle}}{

\code{make_par_list()} defines \mjseqn{\langle\beta\rangle} as a function
of \mjseqn{\mathcal{R}_0} by enforcing the identity

\mjsdeqn{\mathcal{R}_0 = \frac{\langle\beta\rangle N_0}{\gamma + \mu}}

if \code{model = "sir"}, where \mjseqn{\gamma = 1 / (t_\text{lat} + t_\text{inf})},
or the identity

\mjsdeqn{\mathcal{R}_0 = \frac{\langle\beta\rangle N_0}{\gamma + \mu} \frac{\sigma}{\sigma + \mu}}

if \code{model = "seir"}, where \mjseqn{\sigma = 1 / t_\text{lat}}
and \mjseqn{\gamma = 1 / t_\text{inf}}.
}

\subsection{2. Defining \mjseqn{S_0}, \mjseqn{E_0}, and \mjseqn{I_0}}{
\subsection{SIR model}{

If \code{model == "sir"}, then \code{make_par_list()} defines \mjseqn{S_0},
and \mjseqn{I_0} by numerically integrating the system of equations

\mjsdeqn{\begin{align*} \frac{\text{d}S}{\text{d}t} &= \mu_\text{c} N_0 - \beta(t) S I - \mu_\text{c} S \cr \frac{\text{d}I}{\text{d}t} &= \beta(t) S I - \gamma I - \mu_\text{c} I \end{align*}}

with \mjseqn{\gamma = 1 / (t_\text{lat} + t_\text{inf})} and

\mjsdeqn{\beta(t) = \langle\beta\rangle \left\lbrack 1 + \alpha \cos\left(\frac{2 \pi t}{\text{365 days}}\right) \right\rbrack}

between times \mjseqn{t = t_{-(n-1)} = -(n-1) \Delta t}
and \mjseqn{t = 0} years, taking

\mjsdeqn{\begin{bmatrix} S(t_{-(n-1)}) \cr I(t_{-(n-1)}) \end{bmatrix} = N_0 \begin{bmatrix} \frac{1}{\mathcal{R}_0} \cr \big(1 - \frac{1}{\mathcal{R}_0}\big) \frac{\mu}{\gamma + \mu} \end{bmatrix}}

for the initial state. This is the endemic equilibrium of the above system
with constant transmission rate \mjseqn{\beta(t) \equiv \langle\beta\rangle}.
Then \code{make_par_list()} assigns \mjseqn{S_0}, \mjseqn{E_0}, and \mjseqn{I_0}
the value of \mjseqn{S(0)}, \mjseqn{E(0)}, and \mjseqn{I(0)}, respectively.
}

\subsection{SEIR model}{

If \code{model == "seir"}, then \code{make_par_list()} defines \mjseqn{S_0},
\mjseqn{E_0}, and \mjseqn{I_0} by numerically integrating the system
of equations

\mjsdeqn{\begin{align*} \frac{\text{d}S}{\text{d}t} &= \mu_\text{c} N_0 - \beta(t) S I - \mu_\text{c} S \cr \frac{\text{d}E}{\text{d}t} &= \beta(t) S I - \sigma E - \mu_\text{c} E \cr \frac{\text{d}I}{\text{d}t} &= \sigma E - \gamma I - \mu_\text{c} I \end{align*}}

with \mjseqn{\sigma = 1 / t_\text{lat}}, \mjseqn{\gamma = 1 / t_\text{inf}},
and

\mjsdeqn{\beta(t) = \langle\beta\rangle \left\lbrack 1 + \alpha \cos\left(\frac{2 \pi t}{\text{365 days}}\right) \right\rbrack}

between times \mjseqn{t = t_{-(n-1)} = -(n-1) \Delta t}
and \mjseqn{t = 0} years, taking

\mjsdeqn{\begin{bmatrix} S(t_{-(n-1)}) \cr E(t_{-(n-1)}) \cr I(t_{-(n-1)}) \end{bmatrix} = N_0 \begin{bmatrix} \frac{1}{\mathcal{R}_0} \cr \big(1 - \frac{1}{\mathcal{R}_0}\big) \frac{\mu}{\sigma + \mu} \cr \big(1 - \frac{1}{\mathcal{R}_0}\big) \frac{\sigma}{\sigma + \mu} \frac{\mu}{\gamma + \mu} \end{bmatrix}}

for the initial state. This is the endemic equilibrium of the above system
with constant transmission rate \mjseqn{\beta(t) \equiv \langle\beta\rangle}.
Then \code{make_par_list()} assigns \mjseqn{S_0}, \mjseqn{E_0}, and \mjseqn{I_0}
the value of \mjseqn{S(0)}, \mjseqn{E(0)}, and \mjseqn{I(0)}, respectively.
}

}

\subsection{3. Choosing \mjseqn{n}}{

Choosing \mjseqn{n} such that \mjseqn{n \Delta t \sim 1000} years is
typically enough to ensure that \mjseqn{\big(S(0),I(0),E(0)\big)} is
near the attractor of the system, which can be desirable when simulating
epidemic time series data (e.g., using \code{\link[=make_data]{make_data()}}).
}
}

\examples{
# Creates a list for measles in England by default
par_list <- make_par_list()
unlist(par_list)

# Time (rate) parameters should be specified
# in units (per unit) of the observation interval
dt_days <- 7
tlat_days <- 5
tinf_weeks <- 8 / 7
muconst_peryear <- 0.04
par_list <- make_par_list(
  dt_days = dt_days,
  tlat    = tlat_days / dt_days,
  tinf    = tinf_weeks * 7 / dt_days,
  muconst = muconst_peryear * dt_days / 365
)
unlist(par_list)

}
\seealso{
\code{\link[=make_data]{make_data()}}
}
