% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_par_list.R
\name{make_par_list}
\alias{make_par_list}
\title{\loadmathjax
Create a list of parameter values for simulations}
\usage{
make_par_list(
  dt_weeks = 1,
  t0 = 1000 * (365/7)/dt_weeks,
  prep = 1,
  trep = 0,
  k = NULL,
  hatN0 = 1e+06,
  N0 = NULL,
  S0 = NULL,
  I0 = NULL,
  nu = 0.04 * (7/365) * dt_weeks,
  mu = 0.04 * (7/365) * dt_weeks,
  tgen = 13 * (1/7)/dt_weeks,
  Rnaught = 20,
  beta_mean = NULL,
  alpha = 0.08,
  epsilon2 = 0
)
}
\arguments{
\item{dt_weeks}{\mjseqn{\lbrace\,\Delta t\,\rbrace}
Numeric scalar. Observation interval in weeks.}

\item{t0}{\mjseqn{\lbrace\,t_0/\Delta t\,\rbrace}
Numeric scalar. Time of the first observation
in units \mjseqn{\Delta t}.}

\item{prep}{\mjseqn{\lbrace\,p_\text{rep}\,\rbrace}
Numeric scalar. Probability that an infection is reported.}

\item{trep}{\mjseqn{\lbrace\,t_\text{rep}/\Delta t\,\rbrace}
Numeric scalar. Mean time between infection and reporting
in units \mjseqn{\Delta t}.}

\item{k}{\mjseqn{\lbrace\,k\,\rbrace}
Numeric scalar. An optional dispersion parameter.
If specified, then the time between infection and reporting in
units \mjseqn{\Delta t} will be modeled by \code{\link[=make_data]{make_data()}} as a
negative binomially distributed random variable with dispersion
\code{k} and mean \code{trep}. If not, then the delay will be fixed equal
to \code{round(trep)}.}

\item{hatN0}{\mjseqn{\lbrace\,\widehat{N}_0\,\rbrace}
Numeric scalar. Population size at time \mjseqn{t = 0} years.}

\item{N0}{\mjseqn{\lbrace\,N_0\,\rbrace}
Numeric scalar. Population size at time \mjseqn{t = t_0}.
This argument is optional (see Details 2).}

\item{S0}{\mjseqn{\lbrace\,S_0\,\rbrace}
Numeric scalar. Number of susceptibles at time \mjseqn{t = t_0}.
This argument is optional (see Details 2).}

\item{I0}{\mjseqn{\lbrace\,I_0\,\rbrace}
Numeric scalar. Number of infecteds at time \mjseqn{t = t_0}.
This argument is optional (see Details 2).}

\item{nu}{\mjseqn{\lbrace\,\nu \Delta t\,\rbrace}
Numeric scalar. Birth rate expressed per unit \mjseqn{\Delta t}
and relative to \mjseqn{\widehat{N}_0}.}

\item{mu}{\mjseqn{\lbrace\,\mu \Delta t\,\rbrace}
Numeric scalar. Natural mortality rate expressed per unit
\mjseqn{\Delta t} and per capita.}

\item{tgen}{\mjseqn{\lbrace\,t_\text{gen}/\Delta t\,\rbrace}
Numeric scalar. Mean generation interval of the disease
of interest in units \mjseqn{\Delta t}.}

\item{Rnaught}{\mjseqn{\lbrace\,\mathcal{R}_0\,\rbrace}
Numeric scalar. Basic reproduction number of the disease
of interest. If specified (non-\code{NULL}), then \code{beta_mean} is
calculated internally as a function of \code{Rnaught} (see Details 1).}

\item{beta_mean}{\mjseqn{\lbrace\,\langle\beta\rangle \Delta t\,\rbrace}
Numeric scalar. Mean (long-term average) of the seasonally forced
transmission rate \mjseqn{\beta(t)},
expressed per unit \mjseqn{\Delta t} per susceptible per infected.
This argument is ignored if \code{Rnaught} is specified (non-\code{NULL})
(see Details 1).}

\item{alpha}{\mjseqn{\lbrace\,\alpha\,\rbrace}
Numeric scalar. Amplitude of the seasonally forced transmission rate
\mjseqn{\beta(t)}, relative to its mean.}

\item{epsilon2}{\mjseqn{\lbrace\,\epsilon^2\,\rbrace}
Numeric scalar. Variance of the normally distributed phase shift
introduced to the seasonally forced transmission rate \mjseqn{\beta(t)}.}
}
\value{
A list of the arguments of \code{make_par_list()} with values for \code{Rnaught},
\code{beta_mean}, \code{N0}, \code{S0}, and \code{I0} if not supplied in the function call
(see Details 1 and 2).
}
\description{
Creates a list of parameter values to be passed as an argument
of \code{\link[=make_data]{make_data()}} for simulating epidemic time series data.
}
\section{Details}{
\subsection{1. Concordance of \mjseqn{\mathcal{R}_0} and \mjseqn{\langle\beta\rangle}}{

\code{make_par_list()} enforces the identity

\mjsdeqn{\mathcal{R}_0 = \frac{\nu \widehat{N}_0}{\mu} \cdot \frac{\langle\beta\rangle}{\gamma + \mu}\,,}

where \mjseqn{\gamma = 1 / t_\text{gen}}. This means that
it calculates \mjseqn{\langle\beta\rangle} (\code{beta_mean})
as a function of \mjseqn{\mathcal{R}_0} (\code{Rnaught}) unless
\code{Rnaught} is \code{NULL} in the function call, in which case it
calculates \mjseqn{\mathcal{R}_0} as a function of
\mjseqn{\langle\beta\rangle}. Hence argument \code{beta_mean}
is optional if \code{Rnaught} is specified (non-\code{NULL}), but
mandatory otherwise.
}

\subsection{2. Missing \mjseqn{N_0}, \mjseqn{S_0}, or \mjseqn{I_0}}{

If \mjseqn{N_0}, \mjseqn{S_0}, or \mjseqn{I_0} (\code{N0}, \code{S0}, or \code{I0})
is \code{NULL} in the function call, then \code{make_par_list()} uses
\code{\link[deSolve:ode]{deSolve::ode()}} to numerically integrate the system of SIR equations

\mjsdeqn{\begin{align} \frac{\text{d}S}{\text{d}t} &= \nu \widehat{N}_0 - \beta(t) S I - \mu S \cr \frac{\text{d}I}{\text{d}t} &= \beta(t) S I - \gamma I - \mu I \cr \frac{\text{d}R}{\text{d}t} &= \gamma I - \mu R \end{align}}

with \mjseqn{\gamma = 1 / t_\text{gen}} and

\mjsdeqn{\beta(t) = \langle\beta\rangle \left\lbrack 1 + \alpha \cos\left(\frac{2 \pi t}{\text{1 year}}\right) \right\rbrack}

between times \mjseqn{t = 0} years and \mjseqn{t = t_0}, taking

\mjsdeqn{\begin{bmatrix} S(0) \cr I(0) \cr R(0) \end{bmatrix} = \widehat{N}_0 \begin{bmatrix} \frac{1}{\mathcal{R}_0} \cr \big(1 - \frac{1}{\mathcal{R}_0}\big) \frac{\mu}{\gamma + \mu} \cr \big(1 - \frac{1}{\mathcal{R}_0}\big) \frac{\gamma}{\gamma + \mu} \end{bmatrix}}

for the initial state. This is the endemic equilibrium of the above system
with constant transmission rate \mjseqn{\beta(t) \equiv \langle\beta\rangle},
scaled by \mjseqn{\mu/\nu} to enforce
\mjseqn{S(0) + I(0) + R(0) = \widehat{N}_0}
as the initial population size. Then \code{make_par_list()} assigns
\code{N0}, \code{S0}, and \code{I0} (only those \code{NULL} in the function call)
the values

\describe{
\item{\code{N0}}{\mjseqn{S(t_0) + I(t_0) + R(t_0)}}
\item{\code{S0}}{\mjseqn{S(t_0)}}
\item{\code{I0}}{\mjseqn{I(t_0)}}
}
}
}

\examples{
# Creates a reasonable list without user input
par_list <- make_par_list()
unlist(par_list)

# Requires time and rate parameters
# in units of the observation interval
dt_weeks <- 1
nu_peryear <- 0.04
tgen_days <- 13
par_list <- make_par_list(
  dt_weeks = dt_weeks,
  nu       = nu_peryear * (7 / 365) * dt_weeks,
  tgen     = tgen_days * (1 / 7) / dt_weeks
)
unlist(par_list)

}
\seealso{
\code{\link[=make_data]{make_data()}}
}
