% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_par_list.R
\name{make_par_list}
\alias{make_par_list}
\title{Create a list of parameter values}
\usage{
make_par_list(
  dt_weeks = 1,
  t0 = 2000 * (365/7)/dt_weeks,
  prep = 1,
  trep = 0,
  hatN0 = 1e+06,
  N0 = NA,
  S0 = NA,
  I0 = NA,
  nu = 0.04 * (7/365) * dt_weeks,
  mu = 0.04 * (7/365) * dt_weeks,
  tgen = 13 * (1/7)/dt_weeks,
  Rnaught = 20,
  beta_mean = NA,
  alpha = 0.08,
  epsilon = 0
)
}
\arguments{
\item{dt_weeks}{[ \ifelse{latex}{\out{$\Delta t$}}{\ifelse{html}{\out{<i>&Delta;t</i>}}{Dt}} ]
Observation interval in weeks.}

\item{t0}{[ \ifelse{latex}{\out{$t_0$}}{\ifelse{html}{\out{<i>t</i><sub>0</sub>}}{t_0}} ]
Time of the first observation in units
\ifelse{latex}{\out{$\Delta t$}}{\ifelse{html}{\out{<i>&Delta;t</i>}}{Dt}}.}

\item{prep}{[ \ifelse{latex}{\out{$p_\text{rep}$}}{\ifelse{html}{\out{<i>p</i><sub>rep</sub>}}{p_rep}} ]
Case reporting probability.}

\item{trep}{[ \ifelse{latex}{\out{$t_\text{rep}$}}{\ifelse{html}{\out{<i>t</i><sub>rep</sub>}}{t_rep}} ]
Case reporting delay in units
\ifelse{latex}{\out{$\Delta t$}}{\ifelse{html}{\out{<i>&Delta;t</i>}}{Dt}}.}

\item{hatN0}{[ \ifelse{latex}{\out{$\widehat{N}_0$}}{\ifelse{html}{\out{<i>&Ntilde;</i><sub>0</sub>}}{hatN_0}} ]
Population size at time
\ifelse{latex}{\out{$t = 0$}}{\ifelse{html}{\out{<i>t</i> = 0}}{t = 0}}
years (see Details).}

\item{N0}{[ \ifelse{latex}{\out{$N_0$}}{\ifelse{html}{\out{<i>N</i><sub>0</sub>}}{N_0}} ]
Population size at time
\ifelse{latex}{\out{$t = t_0$}}{\ifelse{html}{\out{<i>t</i> = <i>t</i><sub>0</sub>}}{t = t_0}}.
Can be set to \code{NA} (see Details).}

\item{S0}{[ \ifelse{latex}{\out{$S_0$}}{\ifelse{html}{\out{<i>S</i><sub>0</sub>}}{S_0}} ]
Number of susceptibles at time
\ifelse{latex}{\out{$t = t_0$}}{\ifelse{html}{\out{<i>t</i> = <i>t</i><sub>0</sub>}}{t = t_0}}.
Can be set to \code{NA} (see Details).}

\item{I0}{[ \ifelse{latex}{\out{$I_0$}}{\ifelse{html}{\out{<i>I</i><sub>0</sub>}}{I_0}} ]
Number of infecteds at time
\ifelse{latex}{\out{$t = t_0$}}{\ifelse{html}{\out{<i>t</i> = <i>t</i><sub>0</sub>}}{t = t_0}}.
Can be set to \code{NA} (see Details).}

\item{nu}{[ \ifelse{latex}{\out{$\nu_\text{c}$}}{\ifelse{html}{\out{<i>&nu;<sub>c</sub></i>}}{nu_c}} ]
Birth rate expressed per unit
\ifelse{latex}{\out{$\Delta t$}}{\ifelse{html}{\out{<i>&Delta;t</i>}}{Dt}}
and relative to
\ifelse{latex}{\out{$\hat{N}_0$}}{\ifelse{html}{\out{<i>&Ntilde;</i><sub>0</sub>}}{hatN_0}}
(if modeled as constant).}

\item{mu}{[ \ifelse{latex}{\out{$\mu_\text{c}$}}{\ifelse{html}{\out{<i>&mu;</i><sub>c</sub>}}{mu_c}} ]
Natural mortality rate expressed per unit
\ifelse{latex}{\out{$\Delta t$}}{\ifelse{html}{\out{<i>&Delta;t</i>}}{Dt}}
and per capita (if modeled as constant).}

\item{tgen}{[ \ifelse{latex}{\out{$t_\text{gen}$}}{\ifelse{html}{\out{<i>t</i><sub>gen</sub>}}{t_gen}} ]
Mean generation interval of the disease of interest in units
\ifelse{latex}{\out{$\Delta t$}}{\ifelse{html}{\out{<i>&Delta;t</i>}}{Dt}}.}

\item{Rnaught}{[ \ifelse{latex}{\out{$\mathcal{R}_0$}}{\ifelse{html}{\out{<i>&Rscr;</i><sub>0</sub>}}{calR_0}} ]
Basic reproduction number of the disease of interest. Should be set
to \code{NA} when specifying \code{beta_mean} (see Details).}

\item{beta_mean}{[ \ifelse{latex}{\out{$\langle\beta\rangle$}}{\ifelse{html}{\out{&langle;<i>&beta;</i>&rangle;}}{<beta>}} ]
Mean of the seasonally forced transmission rate
\ifelse{latex}{\out{$\beta(t)$}}{\ifelse{html}{\out{<i>&beta;</i>(<i>t</i>)}}{beta(t)}}
expressed per unit
\ifelse{latex}{\out{$\Delta t$}}{\ifelse{html}{\out{<i>&Delta;t</i>}}{Dt}}
per susceptible per infected. Should be set
to \code{NA} when specifying \code{Rnaught} (see Details).}

\item{alpha}{[ \ifelse{latex}{\out{$\alpha$}}{\ifelse{html}{\out{<i>&alpha;</i>}}{alpha}} ]
Amplitude of the seasonally forced transmission rate
\ifelse{latex}{\out{$\beta(t)$}}{\ifelse{html}{\out{<i>&beta;</i>(<i>t</i>)}}{beta(t)}}
relative to the mean.}

\item{epsilon}{[ \ifelse{latex}{\out{$\epsilon$}}{\ifelse{html}{\out{<i>&straightepsilon;</i>}}{epsilon}} ]
Standard deviation of the random phase shift in the seasonally
forced transmission rate
\ifelse{latex}{\out{$\beta(t)$}}{\ifelse{html}{\out{<i>&beta;</i>(<i>t</i>)}}{beta(t)}}.}
}
\value{
A list of the arguments of \code{make_par_list()}, including values for
\code{Rnaught}, \code{beta_mean}, \code{N0}, \code{S0}, and \code{I0} if not defined in the
function call (see Details).
}
\description{
\code{make_par_list()} creates a list of parameter values that may be
passed as an argument of:
\itemize{
\item \code{\link[=make_data]{make_data()}} to simulate time series data.
\item \code{\link[=estimate_beta_FC]{estimate_beta_FC()}}, \code{\link[=estimate_beta_S]{estimate_beta_S()}}, and \code{\link[=estimate_beta_SI]{estimate_beta_SI()}}
to estimate time-varying transmission rates
\ifelse{latex}{\out{$\beta(t)$}}{\ifelse{html}{\out{<i>&beta;</i>(<i>t</i>)}}{beta(t)}}.
\item \code{\link[=ptpi]{ptpi()}} to estimate the initial number of susceptibles
\ifelse{latex}{\out{$S_0$}}{\ifelse{html}{\out{<i>S</i><sub>0</sub>}}{S_0}}.
}

As these methods deal with equally spaced time series data
with fixed observation interval
\ifelse{latex}{\out{$\Delta t$}}{\ifelse{html}{\out{<i>&Delta;t</i>}}{Dt}},
\code{make_par_list()} defines time (or rate) parameters in units
(or per unit)
\ifelse{latex}{\out{$\Delta t$}}{\ifelse{html}{\out{<i>&Delta;t</i>}}{Dt}}.
}
\details{
\code{make_par_list()} enforces the identity

\ifelse{latex}{\out{$\mathcal{R}_0 = \frac{\nu_\text{c} \widehat{N}_0}{\mu_\text{c}} \cdot \frac{\langle\beta\rangle}{\gamma + \mu}$}}{\ifelse{html}{\out{<i>&Rscr;</i><sub>0</sub> = (<i>&nu;</i><sub>c</sub> <i>&Ntilde;</i> / <i>&mu;</i><sub>c</sub>)(&langle;<i>&beta;</i>&rangle; / (<i>&gamma;</i> + <i>&mu;</i><sub>c</sub>))}}{calR_0 = ((nu_c*hatN0)/mu_c)*(<beta>/(gamma + mu))}}

as follows. If exactly one of \code{Rnaught} and \code{beta_mean} is \code{NA} in
in the function call, then that parameter is internally assigned the
value satisfying the identity. If both are \code{NA}, then an error is
thrown. If neither is \code{NA}, then the value of \code{beta_mean} is replaced
with the value satisfying the identity.

If any of \code{N0}, \code{S0}, and \code{I0} is \code{NA} in the function call, then
\code{make_par_list()} numerically integrates the system of SIR equations

\ifelse{latex}{
  \out{
    \begin{array}[rlc]
      S' & = & \nu_\text{c} \widehat{N}_0 - \beta(t) S I - \mu_\text{c} S \\
      I' & = & \beta(t) S I - \gamma I - \mu_\text{c} I \\
      R' & = & \gamma I - \mu_\text{c} R
    \end{array}
  }
}{
  \ifelse{html}{
    \out{
      <i>S</i>&prime; = <i>&nu;</i><sub>c</sub><i>&Ntilde;</i><sub>0</sub> &minus; <i>&beta;</i>(<i>t</i>)<i>SI</i> &minus; <i>&mu;</i><sub>c</sub><i>S</i><br>
      <i>I</i>&prime; = <i>&beta;</i>(<i>t</i>)<i>SI</i> &minus; <i>&gamma;I</i> &minus; <i>&mu;</i><sub>c</sub><i>I</i><br>
      <i>R</i>&prime; = <i>&gamma;I</i> &minus; <i>&mu;</i><sub>c</sub><i>R</i>
    }
  }{
    S' = nu_c*hatN_0 - beta(t)*S*I - mu_c*S \cr
    I' = beta(t)*S*I - gamma*I - mu_c*I \cr
    R' = gamma*I - mu_c*R
  }
}

with
\ifelse{latex}{\out{$\gamma = 1 / t_\text{gen}$}}{\ifelse{html}{\out{<i>&gamma;</i> = 1 / <i>t</i><sub>gen</sub>}}{gamma = 1/t_gen}}
and
\ifelse{latex}{\out{$\beta(t) = \langle\beta\rangle \left(1 + \alpha \cos\left(\frac{2 \pi t}{\text{1 year}}\right)\right)$}}{\ifelse{html}{\out{<i>&beta;</i>(<i>t</i>) = &langle;<i>&beta;</i>&rangle; (1 + <i>&alpha;</i> cos(2<i>&pi;t</i> / (1 year)))}}{beta(t) = <beta>*(1 + alpha*cos(2*pi*t/(1 year)))}}
between times
\ifelse{latex}{\out{$t = 0$}}{\ifelse{html}{\out{<i>t</i> = 0}}{t = 0}}
years and
\ifelse{latex}{\out{$t = t_0$}}{\ifelse{html}{\out{<i>t</i> = <i>t</i><sub>0</sub>}}{t = t_0}},
taking for the initial state
\ifelse{latex}{\out{$(S(0),I(0),R(0))$}}{\ifelse{html}{\out{(<i>S</i>(0),<i>I</i>(0),<i>R</i>(0))}}{(S(0),I(0),R(0))}}
the endemic equilibrium of the system with
\ifelse{latex}{\out{$\beta(t) \equiv \langle\beta\rangle$}}{\ifelse{html}{\out{<i>&beta;</i>(<i>t</i>) &equiv; &langle;<i>&beta;</i>&rangle;}}{beta(t) = <beta>}}
and
\ifelse{latex}{\out{$\nu_\text{c} = \mu_\text{c}$}}{\ifelse{html}{\out{<i>&nu;</i><sub>c</sub> = <i>&mu;</i><sub>c</sub>}}{nu_c = mu_c}}.
Then \code{make_par_list()} defines \code{N0}, \code{S0}, and \code{I0}
(only those that were \code{NA} in the function call) as follows:

\describe{
\item{\code{N0}}{The value of
\ifelse{latex}{\out{$S(t_0)+I(t_0)+R(t_0)$}}{\ifelse{html}{\out{<i>S</i>(<i>t</i><sub>0</sub>)+<i>I</i>(<i>t</i><sub>0</sub>)+<i>R</i>(<i>t</i><sub>0</sub>)}}{S(t_0)+I(t_0)+R(t_0)}}.
}
\item{\code{S0}}{The value of
\ifelse{latex}{\out{$S(t_0)$}}{\ifelse{html}{\out{<i>S</i>(<i>t</i><sub>0</sub>)}}{S(t_0)}}.
}
\item{\code{I0}}{The value of
\ifelse{latex}{\out{$I(t_0)$}}{\ifelse{html}{\out{<i>I</i>(<i>t</i><sub>0</sub>)}}{I(t_0)}}.
}
}
}
\examples{
# All arguments have default values
par_list <- make_par_list()
unlist(par_list)

# Time and rate parameters must be specified
# in terms of the observation interval
dt_weeks <- 1
nu_peryear <- 0.04
tgen_days <- 13
par_list <- make_par_list(
  dt_weeks = dt_weeks,
  nu       = nu_peryear * (7 / 365) * dt_weeks,
  tgen     = tgen_days * (1 / 7) / dt_weeks
)
unlist(par_list)

}
