\name{ptpi}
\alias{ptpi}
\title{Peak to peak iteration}
\description{}
\usage{
ptpi(cases, births, mu,
     start, a = 1L, b = length(cases),
     tol = 1e-06, iter.max = 20L, complete = FALSE)
}
\arguments{
  \item{cases}{}
  \item{births}{}
  \item{mu}{}
  \item{start}{}
  \item{a}{}
  \item{b}{}
  \item{tol}{}
  \item{iter.max}{}
  \item{complete}{}
}
\details{}
\value{}
\seealso{}
\references{}
\examples{}
